### 2024-09-10 #256 @moorepants
Dear @labarba, I can possibly edit this one, but I have one concern. Most of the materials are in Spanish. Do we only accept English materials? I have very basic understanding of Spanish myself, so I can't read such technical writing. The repository says that the materials will have an English translation in the future, but it doesn't seem to be the case presently. How should that be handled?
