+++ paper/paper.md
@@ -34,16 +34,16 @@ bibliography: paper.bib
 lang: "en-GB"
 ---
 
-
 <!--  Notas sobre modificaciones y cambios realizados
 1. single-degree, single degree: ver nota de estilo de PR. También libro Fasano. Lo usaría como "single degree".
 2. "constraint forces", en Fasano, et al aparece como "constraint reaction" y luego lo llaman fuerza (p. 77)
 -->
 
-# Summary
 
+
+# Summary
 We present a code-based undergraduate course on analytical mechanics for engineering students with little to no prior programming knowledge.
-This 16-week flipped classroom [@bishop2013flipped:2013] course provides skills to calculate dynamics and strains of simple mechanical devices, modelled as rigid bodies by solving Euler-Lagrange equations.
+This 16-week flipped classroom [@bishop_flipped_2013:2013] course provides skills to calculate dynamics and strains of simple mechanical devices, modelled as rigid bodies by solving Euler-Lagrange equations.
 <!--  Modificado MAR_20240830
 Each example and practice exercise are solved using computer-based analytical and numerical calculations with the aim to deflect students' focus of attention away from repetitive mathematical tasks towards the physics modelling.
 -->
@@ -58,7 +58,7 @@ Para mejorar el rendimiento académico se necesita aplicar estrategias de enseñ
 De esta manera, el aprendizaje Profundo se encuentra estrechamente relacionado con un tipo de enseñanza basada en su conocimiento previo, la puesta en práctica a través de pruebas de ensayo y error, y la evaluación holística de los contenidos, entre otras cosas (Biggs & Tang, 2011).
 -->
 
-The course addresses specific regional issues faced by third-year Latin American students (mid-career), that by then have learned how to solve ordinary differential equations.
+The course addresses specific regional issues faced by third-year Latin American students (mid-career), who by then have learned how to solve ordinary differential equations.
 Theory and examples exercises, along with the _Python_ code that solves them are presented in _Jupyter notebooks_ run online to avoid installation and hardware requirement issues.
 Currently, the material is available in a GitHub repository in Spanish and has only been partially translated into English.
 <!--
@@ -66,11 +66,12 @@ It is free to use and modify under a Creative Commons licence (CC BY-NC-SA 4.0) 
 -->
 
 
+
 ## Statement of need
-Latin American public universities face two simultaneous constrains: tight budgets and the need to accommodate their classes' schedules to day-working students [@vallejo_google_2022:2022].
+Latin American public universities face two simultaneous constraints: tight budgets and the need to accommodate their classes' schedules to day-working students [@vallejo_google_2022:2022].
 These cash-stripped universities seldom avail computing resources for courses that are not directly related to computer science or programming. 
 Also, as undergraduate programs on engineering at Latin American universities are usually longer than the three-year bachelor's degrees at their Anglo-Saxon counterparts, it is quite common for students to already be part of the labour market while studying. 
-As a result, they have tight schedules and are often unable to attend to university during daytime hours. 
+As a result, they have tight schedules and are often unable to attend university during daytime hours. 
 <!--
 Ojalá fuera cierto. Sinceramiento: ni saben que existen las alternativas de software libre. 
 Authorities of these cash-stripped universities often ask that any solution to introduce students to coding skills should be based upon open-source fee-free resources.
@@ -92,9 +93,10 @@ Unicamp, Brasil, 10 semestres
 https://www.dac.unicamp.br/sistemas/catalogos/grad/catalogo2023/cursos/10g/sugestao.html
 -->
 
-The course presented addresses those issues by providing a free, online, and asynchronous learning environment allowing students to study at their own pace through the flipped classroom approach [@moraros_flipping_2015:2015].
[…]
