### 2024-10-12 6244ba3
Create action.yml
intento de traducción automática
