### 2025-04-12 5c6e272
init translation v2
