### 2025-04-19 c502e00
01 vectorial ready to ask
