+++ 01Vector/exercisesVector.tex
@@ -9,8 +9,11 @@
 
 If you are able to solve these problems on your own, then you can assume that you have the minimum knowledge about these topics. 
 
+\noindent
 The problems marked with (*) have additional difficulties. Don't hesitate about seeking help from teachers and your classmates if you are not able to complete them.
 		
+\noindent
+For each of the following excercises, create a Jupyter notebook with your name in the title, including one or several cells of code interwined with cells showing text indicating the excercise that is being solved.
 
 \begin{enumerate}
 % 	\section*{Vector posición}
@@ -83,10 +86,14 @@ The problems marked with (*) have additional difficulties. Don't hesitate about 
 	\item
 		\begin{minipage}[t][3.5cm]{0.7\textwidth}
 			(*) \textbf{Position as a function of a variable}\\
-			A particle of mass \(m\) is attached to a ring of radius $R$, and therefore its radius measured from the center of the ring is constant. Then it's enough to know the angle $\varphi$ to describe its position. 
+			A particle of mass \(m\) is attached to a ring of radius $R$, and therefore its radius measured from the center of the ring is constant. Then it's enough to know the angle $\varphi$ to describe its position.
 			\begin{enumerate}
-				\item Write it using cartesian coordinates.
-				\item Calculate the velocity of this particle.
+				\item Write it using cartesian coordinates in terms of $R$ and $\varphi$. Recall that $R$ is a constant, it's just a symbol for \verb'SymPy', and $\varphi$ is a variable that depends on time, or dynamic symbol, as it is called in this library. You will need to use trigonometric functions, investigate how these are implemented using this library.
+				\item Calculate the velocity of this particle using \verb'SymPy'.\\
+				Answer:\\ 
+				\(
+					-  R \sin{\left(\varphi \right)} \dot{\varphi}\mathbf{\hat{e}_x} + R \cos{\left(\varphi \right)} \dot{\varphi}\mathbf{\hat{e}_y}
+				\)
 			\end{enumerate}
 		\end{minipage}
 		\begin{minipage}[c][1.5cm][t]{0.15\textwidth}
