### 2025-04-19 e7e57b3
Working on Energy
