+++ 03EulerLagrange/exercisesEulerLagrange.tex
@@ -1,6 +1,8 @@
 \documentclass[11pt, a4paper, twoside]{article}
 
-\input{../figuresLaTeX/introUNLaM}
+% LTeX: language = "es-AR"
+
+\input{../figurasLaTeX/introUNLaM}
 
 
 \begin{document}
@@ -21,40 +23,37 @@ Los problemas marcados con (*) tienen alguna dificultad adicional, no dude en co
 \textbf{Péndulo de punto de suspensión libre y péndulo doble} [Landau \S5 ejs. 1 y 2]\\ 
 Aplique la ecuación de Euler-Lagrange para obtener las ecuaciones de la dinámica de los sistemas:
 \begin{tasks}(3)
-	\task \begin{tikzpicture}[scale= 1.3]
-  	\draw [arrows=-latex] (-1,2) -> (-1,1) node [above=15, right=2] {\(\vec{g}\)}; % g vertical
-		\draw [ultra thick] (-1.5,3) -- (2,3);
-		\fill [pattern = north east lines] (-1.5,3) rectangle (2,3.2); % techo
-		\draw [dashed] (0,3) -- (0,-.25);	% vertical
-		\draw [thick] (0,3) -- +(-60:3) node[midway,above,right=2] {\(\ell\)};	% inclinada +:relativa, -60 grados, longitud 3
-		\shade [ball color=black!80] ($(0,3)+(-60:3)$) circle(0.25) node [] {\color{white} $m$};
-    \draw [arrows=-latex] (0,.4) -> (1.25,.4) node [midway, above] {\( \psi \)}; % desplazamiento horizontal
-		\draw [arrows=-latex] (0,0) arc [start angle=-90, end angle=-65, radius=3] node [below=12, left=8] {\( \varphi \)};
-	\end{tikzpicture}
-	\task \includegraphics[height=0.23\textwidth]{figures/landauS52_fig2}
-	\task \includegraphics[height=0.23\textwidth]{figures/landauS52_fig1}
+	\task \input{figs/pendulum.tikz}
+	\task \includegraphics[height=0.23\textwidth]{landauS52_fig2}
+	\task \includegraphics[height=0.23\textwidth]{landauS52_fig1}
 \end{tasks}
+Resultado 1c:\\
+\(
+\ell_{1} \left(\ell_{1} m_{1} \ddot{\varphi}_{1} + \ell_{1} m_{2} \ddot{\varphi}_{1} + \ell_{2} m_{2} \sin{\left(\varphi_{1} - \varphi_{2} \right)} \dot{\varphi}_{2}^{2} + \ell_{2} m_{2} \cos{\left(\varphi_{1} - \varphi_{2} \right)} \ddot{\varphi}_{2} + g m_{1} \sin{\left(\varphi_{1} \right)} + g m_{2} \sin{\left(\varphi_{1} \right)}\right) = 0\\
+\ell_{2} m_{2} \left(\ell_{1} \sin{\left(\varphi_{1} - \varphi_{2} \right)} \dot{\varphi}_{1}^{2} - \ell_{1} \cos{\left(\varphi_{1} - \varphi_{2} \right)} \ddot{\varphi}_{1} - \ell_{2} \ddot{\varphi}_{2} - g \sin{\left(\varphi_{2} \right)}\right) = 0
+\)
 
 
 \item
 	\begin{minipage}[t][4.5cm]{0.7\textwidth}
 		\textbf{Plano inclinado móvil}\\
-		Un bloque de masa \(m\) está originalmente inmovil sobre un plano de inclinación \(\theta\) que no le presenta fricción y de masa \(M\).
[…]
