+++ 03EulerLagrange/exercisesEulerLagrange.tex
@@ -2,32 +2,30 @@
 
 % LTeX: language = "es-AR"
 
-\input{../figurasLaTeX/introUNLaM}
+\input{../figuresLaTeX/introUNLaM}
+
 
 
 \begin{document}
 \begin{center}
-  % \textsc{\large Mecánica general}\\
-  \textsc{\large Ecuación de Euler-Lagrange}
+  \textsc{\large Euler-Lagrange equation}
 \end{center}
 
 \noindent
-%De poder resolver estos problemas en forma autónoma puede asumir que adquirió los conocimientos mínimos sobre los temas abordados en la semana.
-%No dude en consultar a docentes y compañeros si no puede terminarlos.
-Los problemas marcados con (*) tienen alguna dificultad adicional, no dude en consultar.
+Exercises marked with (*) have extra difficulty, don't hesitate to ask for help.
+
 \begin{enumerate} 
 
-
 \item
-\textbf{Péndulo rígido ideal} [Marion (english) ex. 7.2] \\
-\textbf{Péndulo de punto de suspensión libre y péndulo doble} [Landau \S5 ejs. 1 y 2]\\ 
-Aplique la ecuación de Euler-Lagrange para obtener las ecuaciones de la dinámica de los sistemas:
+\textbf{Ideal physical pendulum} [Marion ex. 7.2] \\ 
+\textbf{Pendulum with free point of support and double pendulum} [Landau \S5 exs. 1 y 2]\\
+Apply the Euler-Lagrange equation to obtain the equations that describe the dynamics of the following systems:
 \begin{tasks}(3)
-	\task \input{figs/pendulum.tikz}
-	\task \includegraphics[height=0.23\textwidth]{landauS52_fig2}
-	\task \includegraphics[height=0.23\textwidth]{landauS52_fig1}
+	\task \input{figures/pendulum.tikz}
+	\task \includegraphics[height=0.23\textwidth]{figures/landauS52_fig2}
+	\task \includegraphics[height=0.23\textwidth]{figures/landauS52_fig1}
 \end{tasks}
-Resultado 1c:\\
+Result 1c:\\
 \(
 \ell_{1} \left(\ell_{1} m_{1} \ddot{\varphi}_{1} + \ell_{1} m_{2} \ddot{\varphi}_{1} + \ell_{2} m_{2} \sin{\left(\varphi_{1} - \varphi_{2} \right)} \dot{\varphi}_{2}^{2} + \ell_{2} m_{2} \cos{\left(\varphi_{1} - \varphi_{2} \right)} \ddot{\varphi}_{2} + g m_{1} \sin{\left(\varphi_{1} \right)} + g m_{2} \sin{\left(\varphi_{1} \right)}\right) = 0\\
 \ell_{2} m_{2} \left(\ell_{1} \sin{\left(\varphi_{1} - \varphi_{2} \right)} \dot{\varphi}_{1}^{2} - \ell_{1} \cos{\left(\varphi_{1} - \varphi_{2} \right)} \ddot{\varphi}_{1} - \ell_{2} \ddot{\varphi}_{2} - g \sin{\left(\varphi_{2} \right)}\right) = 0
@@ -36,41 +34,36 @@ Resultado 1c:\\
 
 \item
 	\begin{minipage}[t][4.5cm]{0.7\textwidth}
-		\textbf{Plano inclinado móvil}\\
-		Un bloque de masa \(m_1\) está originalmente inmovil sobre un plano de inclinación \(\theta\) que no le presenta fricción y de masa \(M\).
-		Este último puede deslizar sobre la superficie horizontal que tampoco le presenta fricción alguna.
[…]
