### 2025-04-19 39f8904
Set up for starting 04. Paused until all previous images are fixed (embedded).
