+++ 04Constraints/exercisesConstraints.tex
@@ -1,6 +1,6 @@
 \documentclass[11pt, spanish, a4paper, twoside]{article}
 
-\input{../figuresLaTeX/introUNLaM}
+\input{../figurasLaTeX/introUNLaM}
 
 
 \begin{document}
@@ -13,127 +13,141 @@
 %De poder resolver estos problemas en forma autónoma puede asumir que adquirió los conocimientos mínimos sobre los temas abordados en la semana.
 %No dude en consultar a docentes y compañeros si no puede terminarlos.
 Los problemas marcados con (*) tienen alguna dificultad adicional, no dude en consultar.
+
 \begin{enumerate}
 
-
 \item 
-\begin{minipage}[t][3.5cm]{0.68\textwidth}
-	\textbf{Máquina de Atwood simple}\\
-	Obtenga a partir de la ecuación de Euler-Lagrange la aceleración que presentan las pesas de masas \(m_1\) y \(m_2\) que cuelgan de una cuerda de longitud \(\ell\) que pasa por sobre una polea de radio \(R_p\) y masa \(m_p\).
-	\begin{enumerate}
-		\item Resuelva el caso en que se considera \(m_p\) irrelevante.
-		\item Resuelva ahora considerando \(m_p\), y que la polea presenta una sección cilíndrica.
-			El momento de inercia de tal cilindro de masa \(m\) ante rotaciones en torno a su eje de simetría longitudinal es \((m/2) R^2\).
-	\end{enumerate}
-\end{minipage}
-\begin{minipage}[c][0.5cm][t]{0.3\textwidth}
-	% \hspace{0.5cm}
-	\includegraphics[width=0.75\textwidth]{figures/marion_fig2_11a}
-\end{minipage}
-
-
-\item 
-	\begin{minipage}[t][3.5cm]{0.8\textwidth}
-		\textbf{Aro y polea}\\
-		Una pesa de masa \(m\) pende de una sección de cuerda de longitud \(\ell_{pesa}\) que sobresale a la derecha de una polea de radio \(R_{polea}\) y masa \(m_{poleo}\).
-		La cuerda tiene un longitud total \(\ell\), su  masa es despreciable y gira solidaria con la una polea.
-
-		El otro extremo se ata con un nudo de masa \(M_{nude}\) a un aro de masa \(m_{aro}\), enrollándose parcialmente en torno a éste.
-		El centro de la polea está a una altura \(h\) por sobre el del aro de radio \(R_{aro}\) que como puede rotar libremente presenta un momento de inercia \(m_{aro} R_{aro}^2\).
-		
-		Denomine el ángulo desde el centro hasta el nudo medido desde la horizontal con \(\theta\).
-		Recuerde que el sentido positivo para un ángulo es el antihorario. 
+	\begin{minipage}[t][3cm]{0.72\textwidth}
+		\textbf{Máquina de Atwood simple}\\
+		Una cuerda de longitud \(\ell\) pasa sobre una polea de radio \(R_p\) y masa \(m_p\).
+		Calcule la aceleración de las pesas que cuelgan de sus extremos.
 	\end{minipage}
-	\begin{minipage}[c][0cm][t]{0.2\textwidth}
+	\begin{minipage}[c][3cm][t]{0.2\textwidth}
 		% \hspace{0.5cm}
[…]
