+++ 04Constraints/exercisesConstraints.tex
@@ -1,6 +1,6 @@
 \documentclass[11pt, spanish, a4paper, twoside]{article}
 
-\input{../figurasLaTeX/introUNLaM}
+\input{../figuresLaTeX/introUNLaM}
 
 
 \begin{document}
@@ -24,9 +24,7 @@ Los problemas marcados con (*) tienen alguna dificultad adicional, no dude en co
 	\end{minipage}
 	\begin{minipage}[c][3cm][t]{0.2\textwidth}
 		% \hspace{0.5cm}
-		\input{figs/atwood.tikz}
-		% \includegraphics[width=0.75\textwidth]{marion_fig2_11a}
-		% \includegraphics[width=0.75\textwidth]{marion_fig2_1a}
+		\input{figures/atwood.tikz}
 	\end{minipage}
 		\begin{enumerate}
 			\item La cuerda es inextensible, por lo que establece una relación entre \(y_1\) e \(y_2\).
@@ -47,9 +45,7 @@ Los problemas marcados con (*) tienen alguna dificultad adicional, no dude en co
 		La barra establece un vínculo entre las respectivas coordenadas que definen sus posiciones, \(x\) e \(y\).
 	\end{minipage}
 	\begin{minipage}[c][1cm][t]{0.25\textwidth}
-		\input{figs/slidingWeights.tikz}
-		% \hspace{0.5cm}
-		% \includegraphics[width=0.75\textwidth]{fcen1-004}
+		\input{figures/slidingWeights.tikz}
 	\end{minipage}
 	\begin{enumerate}
 		\item Use la función de vínculo para expresar las posiciones solo en función de \(y\).
@@ -85,7 +81,7 @@ Los problemas marcados con (*) tienen alguna dificultad adicional, no dude en co
 		Quedan el apartamiento con la horizontal de la pesa, \(y_{pesa}\), y la extensión del arco enrollado, \(\theta\), como las coordenadas generalizadas que estarán ligadas por la cuerda de longitud \(\ell\).
 	\end{minipage}
 	\begin{minipage}[c][1.5cm][t]{0.2\textwidth}
-		\input{figs/ringPulley.tikz}
+		\input{figures/ringPulley.tikz}
 	\end{minipage}
 	%\begin{minipage}[c][0cm][t]{0.2\textwidth}
 		%\hspace{0.5cm}
@@ -116,8 +112,7 @@ Los problemas marcados con (*) tienen alguna dificultad adicional, no dude en co
 		\end{enumerate}
 	\end{minipage}
 	\begin{minipage}[c][0cm][t]{0.3\textwidth}
-		\input{figs/compoundAtwood.tikz}
-		%\includegraphics[width=\textwidth]{marion_fig7_6}
+		\input{figures/compoundAtwood.tikz}
 	\end{minipage}
 
 		\begin{enumerate}
