+++ 04Constraints/exercisesConstraints.tex
@@ -1,114 +1,89 @@
-\documentclass[11pt, spanish, a4paper, twoside]{article}
+\documentclass[11pt, a4paper, twoside]{article}
 
 \input{../figuresLaTeX/introUNLaM}
 
 
 \begin{document}
 \begin{center}
-  % \textsc{\large Mecánica general}\\
-  \textsc{\large Ligaduras}
+  \textsc{\large Constraints}
 \end{center}
 
 \noindent
-%De poder resolver estos problemas en forma autónoma puede asumir que adquirió los conocimientos mínimos sobre los temas abordados en la semana.
-%No dude en consultar a docentes y compañeros si no puede terminarlos.
-Los problemas marcados con (*) tienen alguna dificultad adicional, no dude en consultar.
+Exercises marked with (*) have extra difficulty, don't hesitate to ask for help.
 
 \begin{enumerate}
 
 \item 
 	\begin{minipage}[t][3cm]{0.72\textwidth}
-		\textbf{Máquina de Atwood simple}\\
-		Una cuerda de longitud \(\ell\) pasa sobre una polea de radio \(R_p\) y masa \(m_p\).
-		Calcule la aceleración de las pesas que cuelgan de sus extremos.
+		\textbf{Atwood machine}\\
+		The figure shows a string of length \(\ell\) and a pulley of radius \(R_p\) and mass \(m_p\).
+		Find the acceleration of the masses attached at each end of the string.
 	\end{minipage}
 	\begin{minipage}[c][3cm][t]{0.2\textwidth}
-		% \hspace{0.5cm}
 		\input{figures/atwood.tikz}
 	\end{minipage}
 		\begin{enumerate}
-			\item La cuerda es inextensible, por lo que establece una relación entre \(y_1\) e \(y_2\).
-			Escriba esta función de vínculo.
-			\item Si la cuerda desliza sin rozamiento sobre la polea, esta última no se mueve.
-			Calcule la ecuación de Euler-Lagrange para \(y_1\) usando el vínculo del punto anterior y escriba la aceleración de las pesas.
-			\item Lo usual es que la cuerda no deslice y gire solidaria a la polea.
-			Esto vincula \(\theta\) al desplazamiento de la cuerda.
-			Con tal vínculo exprese en función de \(\dot{y}_1\) la energía cinética de rotación de la polea, que modelada como un cilindro homogéneo que en torno a su eje presenta un momento de inercia \((m/2) R^2\).
-			\item Use la ecuación de Euler-Lagrange para \(y_1\) para escribir la aceleración de las pesas en este caso.
+			\item The string is inextensible, so it establishes a relation between \(y_1\) and \(y_2\).
+			Write the function for this constraint.
+			\item If the string slides over the pulley without friction, the pulley will not move.
+			Write the Euler-Lagrange equation for \(y_1\) using the constraint from the previous item and write the masses' acceleration.
+			\item Usually, the string won't slide and the pulley will rotate.
+			This constraint adds a relation between \(\theta\) and the displacement of the string.
+			Using that constraint, write the pulley's rotational kinetic energy in terms of \(\dot{y}_1\), modeling the pulley as an homogeneous cylinder with a moment of inertia of \((m/2) R^2\).
[…]
