### 2025-04-20 ab36024
04 working on constraint
