+++ 04Constraints/exercisesConstraints.tex
@@ -24,7 +24,7 @@ Exercises marked with (*) have extra difficulty, don't hesitate to ask for help.
 	\end{minipage}
 		\begin{enumerate}
 			\item The string is inextensible, so it establishes a relation between \(y_1\) and \(y_2\).
-			Write the function for this constraint.
+			Write the equation for this constraint.
 			\item If the string slides over the pulley without friction, the pulley will not move.
 			Write the Euler-Lagrange equation for \(y_1\) using the constraint from the previous item and write the masses' acceleration.
 			\item Usually, the string won't slide and the pulley will rotate.
@@ -44,7 +44,7 @@ Exercises marked with (*) have extra difficulty, don't hesitate to ask for help.
 		\input{figures/slidingWeights.tikz}
 	\end{minipage}
 	\begin{enumerate}
-		\item Use the constraint function to express both positions only in terms of \(y\).
+		\item Use the constraint equation to express both positions only in terms of \(y\).
 		\item Calculate the acceleration of \(m_2\).\\
 		Result:
 			$\ddot{y} = \frac{- \ell^{2} m_{1} y \dot{y}^{2} + g m_{2} \left(\ell^{2} - y^{2}\right)^{2}}{\ell^{4} m_{2} + \ell^{2} m_{1} y^{2} - 2 \ell^{2} m_{2} y^{2} - m_{1} y^{4} + m_{2} y^{4}}$
@@ -81,8 +81,8 @@ Exercises marked with (*) have extra difficulty, don't hesitate to ask for help.
 		\textbf{Double Atwood machine} [Marion ex. 7.8]\\ 
 		\begin{enumerate}
 			\item Write the positions for the three hanging objects and that of the lower pulley in terms of the generalized coordinates shown in the figure: \(y_i\) with \(i = 1,2,3,p\). 
-			\item Write the functions for the constraints provided by both strings.
-			\item Use the constraint functions to express all positions in terms of \(y_1\) and \(y_2\).
+			\item Write the equations for the constraints provided by both strings.
+			\item Use the constraint equations to express all positions in terms of \(y_1\) and \(y_2\).
 			\item The strings don't slide over the pulleys. This is an extra constraint that must be used to relate \(y_i\) and \(\theta_i\).
 		\end{enumerate}
 	\end{minipage}
