### 2025-06-02 #256 @emckiernan
Hi, @nicoguaro! I'm writing as the editor of this submission, and reaching out because you previously signed up to review for JOSE. Would you like to contribute a review for this submission? Please let me know if you're available and willing. Thanks in advance for considering!
