### 2025-06-03 #256 @nicoguaro
@emckiernan, I am also an editor of JOSE and have one JOSS review pending. If you can wait around a month for me to start, I could accept.
