### 2025-06-03 #256 @emckiernan
Hi, @richmf, thanks for accepting my invitation over email! I'll tag you below to formally assign you as a reviewer. The materials to review are in the repository linked at the top of this page. And here are the reviewer guidelines: https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html. If you have any questions, please just let me know. Thanks again!
