### 2025-06-05 #256 @bettachini
> trans
Hi everyone. We, the co-authors, want to thank you for your effort in finding bilingual reviewers.
A few months ago we wanted to address the inconvenience to the reviewers of not having an English translation. Sadly, we hadn't advanced a lot towards that goal. You can find the contents of the first two directories fully translated here:
https://github.com/unlam/ComputationalAnalyticalMechanics
