### 2025-06-05 #256 @bettachini
> [@bettachini](https://github.com/bettachini) – your materials are both in English and Spanish. The challenge is thus that we need two reviewers who are also bilingual. If you have any recommended reviewers, please comment here (but without @-mentioning them, please).

A fellow mechanical engineer, from an institution neither of the co-authors works at, said he could review the article. How should we send you his contact details?
