### 2025-06-09 #256 @emckiernan
> Hi, [@dpshelio](https://github.com/dpshelio)! I'm writing as the editor of this submission, and reaching out because you previously signed up to review for JOSE. Would you like to contribute a review for this submission? Please let me know if you're available and willing. Thanks in advance for considering!

Hi, @dpshelio, just checking back in. Might you be able to contribute a review for this?
