### 2025-06-09 #256 @emckiernan
> A fellow mechanical engineer, from an institution neither of the co-authors works at, said he could review the article. How should we send you his contact details?

Hi, @bettachini, just to update you...I sent the invite on Thursday, but waiting to hear back.
