### 2025-06-16 #256 @emckiernan
> Hi, [@bettachini](https://github.com/bettachini), just to update you...I sent the invite on Thursday, but waiting to hear back.

Hi, @bettachini, checking in again...have you heard anything from the professor you identified as a potential reviewer? I sent the invite by email but haven't heard back. Maybe it went to his spam? Would be great to know if he is still willing.
