+++ 04Constraints/figures/slidingWeights.tex
@@ -1,0 +1,11 @@
+\documentclass[border=3.14mm]{standalone}
+
+\usepackage{tikz}
+\usetikzlibrary{decorations.pathmorphing} 
+\usetikzlibrary{arrows.meta}
+\usetikzlibrary{patterns}
+\usetikzlibrary{calc}
+ 
+\begin{document}
+\input{slidingWeights.tikz}
+\end{document}
