+++ translation.txt
@@ -1,0 +1,218 @@
+
+tags/202503
+
+TODO:
+- Redo some figures, that look like screenshots, if there are copyright issues.
+- Check the year in the headers and footnotes. Some notebooks say 2022 in English and 2024 in Spanish.
+
+
+- Translate imagesNotebooks.ipynb
+
+
+
+
+== 01Vectorial -> 01Vector 
+
+cinemáticaVectorial.ipynb  ->  vectorKinematics.ipynb   Ready
+entregaEjercicios.ipynb    ->  ?                        Ask.
+estaAsignatura.ipynb       ->  thisCourse.ipynb         Ready
+guíaVectorial.tex          ->  exercisesVector.tex      Ready.
+?                          ->  MarkdownLaTeX.ipynb      Ask.
+?                          ->  IntroPython.ipynb        Ask.
+
+
+== 02Energía -> 02Energy
+
+apunteLanczos.pdf                  ->  ?                                    Ask.
+energíaCinéticaTraslación.ipynb    ->  kineticEnergy.ipynb                  Ready.
+energíaPotencialGravitatoria.ipynb ->  gravitationalPotentialEnergy.ipynb   Ask. (1)
+guíaEnergía.tex                    ->  exercisesEnergy.tex                  Ask. (2)
+?                                  ->  p02e04_help.ipynb                    Ask. Keep or move to pdf?
+
+
+== 03EulerLagrange -> 03EulerLagrange
+
+apunteLanczos.tex                              ->  ?                              Tranlsating (10)
+energíaPotencialElástica_eulerLagrange.ipynb   ->  elasticPotentialEnergy.ipynb   Ask. (3)
+guíaEulerLagrange.tex                          ->  exercisesEulerLagrange.tex     Ready (4, ref.)
+partesRotantes_eulerLagrange.ipynb             ->  rotatingParts.ipynb            Ready
+plantilla_eulerLagrange.ipynb                  ->  euler_lagrange_base.ipynb      Ready
+péndulos_eulerLagrange.ipynb                   ->  eulerLagrangePendulums.ipynb   Ask. (5)
+
+
+== 04Ligaduras  ->  04Constraints
+
+atwood_Ligaduras.ipynb            ->  atwoodConstraint.ipynb        Ready (7)
+guíaLigaduras.tex                 ->  exercisesConstraints.tex      Ask (almost ready) (6)
+ligadurasFunción.ipynb            ->  constraintsFunction.ipynb     Ready (8)
+resoluciónSistemasLineales.ipynb  ->  solvingLinearSystems.ipynb    Ready (9)
+figures\ringPulley.tikz: changed labels.
+
+
+== 05Simulación  ->  05Simulation
+
+atwoodSimulación_Resuelto.ipynb         ->  atwoodSimulationSolved.ipynb
+guíaSimulación.tex                      ->  exercisesSimulation.tex
+pénduloLibre_Simulación_Resuelto.ipynb  ->  pendulumFreeSimulSolved.ipynb
+
+== 06FuerzasLigadura  ->  06ForcesConstraint
+
+fuerzasLigadura.ipynb               ->  forcesConstraint.ipynb
+guíaFuerzasLigadura.tex             ->  exercisesConstraintForces.tex
+pénduloIdeal_FuerzasLigadura.ipynb  ->  constraintIdealPendulum.ipynb
+rodadura_FuerzasLigadura.ipynb      ->  rollingForcesConstraint.ipynb
+sistemasNoHolónomos.ipynb           ->  ?
+
+
+== 07NoConservativas  ->  07NonConservative
+
+cilíndrosSolidarios_noConservativas_ayuda.ipynb  ->  linkedCylinders_Help.ipynb
[…]
