+++ translation.txt
@@ -14,35 +14,34 @@ TODO:
 == 01Vectorial -> 01Vector 
 
 cinemáticaVectorial.ipynb  ->  vectorKinematics.ipynb   Ready
-entregaEjercicios.ipynb    ->  ?                        Ask.
+entregaEjercicios.ipynb    ->  ?                        Ask.  Borrar.
 estaAsignatura.ipynb       ->  thisCourse.ipynb         Ready
 guíaVectorial.tex          ->  exercisesVector.tex      Ready.
-?                          ->  MarkdownLaTeX.ipynb      Ask.
-?                          ->  IntroPython.ipynb        Ask.
+?                          ->  MarkdownLaTeX.ipynb      Ask.  Borrar
+?                          ->  IntroPython.ipynb        Ask.  Borrar
 
 
 == 02Energía -> 02Energy
 
-apunteLanczos.pdf                  ->  ?                                    Ask.
+apunteLanczos.pdf                  ->  notesLanczos.pdf                     Traduciendo Victor, Falta el archivo.
 energíaCinéticaTraslación.ipynb    ->  kineticEnergy.ipynb                  Ready.
-energíaPotencialGravitatoria.ipynb ->  gravitationalPotentialEnergy.ipynb   Ask. (1)
-guíaEnergía.tex                    ->  exercisesEnergy.tex                  Ask. (2)
-?                                  ->  p02e04_help.ipynb                    Ask. Keep or move to pdf?
+energíaPotencialGravitatoria.ipynb ->  gravitationalPotentialEnergy.ipynb   Ready.
+guíaEnergía.tex                    ->  exercisesEnergy.tex                  Ready.
 
 
 == 03EulerLagrange -> 03EulerLagrange
 
-apunteLanczos.tex                              ->  ?                              Tranlsating (10)
-energíaPotencialElástica_eulerLagrange.ipynb   ->  elasticPotentialEnergy.ipynb   Ask. (3)
+apunteLanczos.tex                              ->  notesLanczos.tex               Translating Victor (10)
+energíaPotencialElástica_eulerLagrange.ipynb   ->  elasticPotentialEnergy.ipynb   Ready.
 guíaEulerLagrange.tex                          ->  exercisesEulerLagrange.tex     Ready (4, ref.)
 partesRotantes_eulerLagrange.ipynb             ->  rotatingParts.ipynb            Ready
 plantilla_eulerLagrange.ipynb                  ->  euler_lagrange_base.ipynb      Ready
-péndulos_eulerLagrange.ipynb                   ->  eulerLagrangePendulums.ipynb   Ask. (5)
+péndulos_eulerLagrange.ipynb                   ->  eulerLagrangePendulums.ipynb   Ready
 
 
 == 04Ligaduras  ->  04Constraints
 
-atwood_Ligaduras.ipynb            ->  atwoodConstraint.ipynb        Ready (7)
+atwood_Ligaduras.ipynb            ->  atwoodConstraint.ipynb        Ready (Test that everything works)
 guíaLigaduras.tex                 ->  exercisesConstraints.tex      Ask (almost ready) (6)
 ligadurasFunción.ipynb            ->  constraintsFunction.ipynb     Ready (8)
 resoluciónSistemasLineales.ipynb  ->  solvingLinearSystems.ipynb    Ready (9)
@@ -51,7 +50,7 @@ figures\ringPulley.tikz: changed labels.
 
 == 05Simulación  ->  05Simulation
 
-atwoodSimulación_Resuelto.ipynb         ->  atwoodSimulationSolved.ipynb
+atwoodSimulación_Resuelto.ipynb         ->  atwoodSimulationSolved.ipynb  (ver repo actual en español)
[…]
