+++ 02Energy/exercisesEnergy.tex
@@ -71,7 +71,7 @@ Exercises marked with (*) have extra difficulty, don't hesitate to ask for help.
 
 
 \item
-	\begin{minipage}[t][4.5cm]{0.65\textwidth}
+	\begin{minipage}[t][7cm]{0.65\textwidth}
 		(*) \textbf{Coupled weights rotating about a vertical axis} [Landau \S5 ex. 4]\\
 		Particle with mass $m_2$ moves on a vertical axis and the whole system rotates about this axis with a constant angular velocity $\Omega$. This particle is linked to two particles of mass $m_1$ through bars of length $a$ and negligible mass, and at the same time, these particles are linked to the fixed point $A$ through identical bars, forming the variable angle $\theta$ with respect to the vertical axis.
 		Calculate the kinetic energy of each of the three particles and find a compact expression of the kinetic energy of the whole system. Do the same for the potential energy.
@@ -85,6 +85,21 @@ Exercises marked with (*) have extra difficulty, don't hesitate to ask for help.
 	\begin{minipage}[c][1cm][t]{0.35\textwidth}
 		\includegraphics[width=0.75\textwidth]{figures/landauS52_fig4.png}
 	\end{minipage}
+
+	\begin{center}
+		\includegraphics[width=0.35\textwidth]{figures/rotantesAcopladas1.png}
+		\includegraphics[width=0.35\textwidth]{figures/rotantesAcopladas2.png}
+	\end{center}
+
+	In these figures, the top sphere is the fixed point \(A\). Everything revolves around the pink axis with constant angular speed $\Omega$. Therefore, the particles on each side, of mass $m_1$, rotate entering and leaving the plane shown in the first image. This is equivalent to thinking that the lightblue plane rotates around the pink axis.
+
+	The piece of mass $m_2$ is a sliding hub that moves vertically without friction. 
+	
+	\begin{center}
+		\includegraphics[width=0.35\textwidth]{figures/rotantesAcopladas3.png}
+		\includegraphics[width=0.35\textwidth]{figures/rotantesAcopladas4.png}
+	\end{center}
+	
 
 
 \end{enumerate}
