+++ 03EulerLagrange/exercisesEulerLagrange.tex
@@ -1,7 +1,5 @@
 \documentclass[11pt, a4paper, twoside]{article}
 
-% LTeX: language = "es-AR"
-
 \input{../figuresLaTeX/introUNLaM}
 
 
