+++ 05Simulation/exercisesSimulation.tex
@@ -1,4 +1,4 @@
-\documentclass[11pt, spanish, a4paper, twoside]{article}
+\documentclass[11pt, a4paper, twoside]{article}
 
 \input{../figuresLaTeX/introUNLaM}
 
@@ -6,16 +6,13 @@
 \begin{document}
 \begin{center}
   % \textsc{\large Mecánica general}\\
-  \textsc{\large Simulación de la dinámica | Resolución numérica de la ecuación de Euler-Lagrange}
+  \textsc{\large Simulation | Numerical solutions for the Euler-Lagrange equation}
 \end{center}
 
 \noindent
-%De poder resolver estos problemas en forma autónoma puede asumir que adquirió los conocimientos mínimos sobre los temas abordados en la semana.
-%No dude en consultar a docentes y compañeros si no puede terminarlos.
-En los siguientes problemas resolverá numericamente cada ecuación de Euler-Lagrange que corresponda a cada coordenada generalizada.
-Graficando tales soluciones, en el rango de tiempos y con las condiciones iniciales indicadas, estará simulando la dinámica de tales sistemas.\\
-La aceleracion gravitatoria tiene por magnitud \(|\vec{g}| = \SI{9.81}{\metre\per\second\squared}\).\\
-Los problemas marcados con (*) tienen alguna dificultad adicional, no dude en consultar.
+In the following exercises, you will solve numerically the Euler-Lagrange equation for each generalized coordinate. Plotting these solutions, using the given initial conditions and within the given time ranges, you will be simulating the dynamics of these systems.\\
+Use \(|\vec{g}| = \SI{9.81}{\metre\per\second\squared}\) for the magnitude of the acceleration due to gravity.\\
+Exercises marked with (*) have extra difficulty, don't hesitate to ask for help.
 
 
 \begin{enumerate}
@@ -23,14 +20,14 @@ Los problemas marcados con (*) tienen alguna dificultad adicional, no dude en co
 
 \item 
 \begin{minipage}[t][2.5cm]{0.7\textwidth}
-\textbf{Máquina de Atwood simple}\\
-Rango de tiempo \(t = \SIrange{0}{10}{\second}\).
-Parámetros físicos y condiciones iniciales:\\
-\(\ell_\mathrm{cuerda} > \SI{150}{\metre}\), 
-\(R_{\mathrm{polea}\,1} = \SI{0.5}{\metre}\), \\ 
+\textbf{Atwood machine}\\
+Time from \(t = \SI{0}{\second}\) to \(t = \SI{10}{\second}\).
+Parameters and initial conditions:\\
+\(\ell_\mathrm{rope} > \SI{150}{\metre}\), 
+\(R_{\mathrm{pulley}\,1} = \SI{0.5}{\metre}\), \\ 
 \(m_1 = \SI{8}{\kilo\gram}\), 
 \(m_2 = \SI{1}{\kilo\gram}\), 
-\(M_\mathrm{polea} = \SI{4}{\kilo\gram}\), \\
+\(M_\mathrm{pulley} = \SI{4}{\kilo\gram}\), \\
 \(x(t=0) = \SI{25}{\metre}\), 
 \(\dot{x}(t=0) = -\SI{10}{\metre\per\second}\).
 \end{minipage}
[…]
