### 2025-06-23 40cea11
Exercises - simulation - figures updated to current version.
