### 2025-06-23 b4bb2b6
Atwood: Translating. Exercises: updated.
