+++ 05Simulation/exercisesSimulation.tex
@@ -28,8 +28,8 @@ Parameters and initial conditions:\\
 \(m_1 = \SI{8}{\kilo\gram}\), 
 \(m_2 = \SI{1}{\kilo\gram}\), 
 \(m_p = \SI{4}{\kilo\gram}\), \\
-\(x(t=0) = \SI{25}{\metre}\), 
-\(\dot{x}(t=0) = -\SI{10}{\metre\per\second}\).
+\(x_1(t=0) = \SI{25}{\metre}\), 
+\(\dot{x}_1(t=0) = -\SI{10}{\metre\per\second}\).
 \end{minipage}
 \begin{minipage}[c][2cm][t]{0.3\textwidth}
 	%\hspace{0.5cm}
