### 2025-06-24 #256 @bettachini
Hi @emckiernan. I'd just been informed that Ferrari was unable to find anything on the subject on his e-mail. I'll send you again his address to your google e-mail. Sorry for the delay, but I do not know him personally and communication with him is through a third party.
