### 2025-06-24 c6d2e50
Almost ready, check 2nd reference in Runge-Kutta.
