+++ translation.txt
@@ -49,7 +49,7 @@ figures\ringPulley.tikz: changed labels.
 
 atwoodSimulación_Resuelto.ipynb         ->  atwoodSimulationSolved.ipynb  (11) Almost ready
 guíaSimulación.tex                      ->  exercisesSimulation.tex        Ready
-pénduloLibre_Simulación_Resuelto.ipynb  ->  pendulumFreeSimulSolved.ipynb
+pénduloLibre_Simulación_Resuelto.ipynb  ->  pendulumFreeSimulSolved.ipynb  Ready (12)
 
 == 06FuerzasLigadura  ->  06ForcesConstraint
 
@@ -151,4 +151,7 @@ Dice que usaremos 3 métodos pero solo se muestran 2 en se notebook.
 (11)
 - The second reference in Runge-Kutta is in Spanish. Remove it? Leave it? Find another one?
 - Spanish: fix constaste
+
+(12)
+Translated from current version. The version in the release is outdated.
 
