+++ 06ForcesConstraint/exercisesConstraintForces.tex
@@ -1,30 +1,29 @@
-\documentclass[11pt, spanish, a4paper, twoside]{article}
+\documentclass[11pt, a4paper, twoside]{article}
 
 \input{../figuresLaTeX/introUNLaM}
 
 \begin{document}
 \begin{center}
-  % \textsc{\large Mecánica general}\\
-  \textsc{\large Fuerzas de ligadura | Multiplicadores de Lagrange} 
+  \textsc{\large Constraint forces | Lagrange multipliers} 
 \end{center}
 
 \begin{enumerate}
 
 	\item
 	\begin{minipage}[t][1.1cm]{0.75\textwidth}
-		\textbf{Péndulo rígido ideal}\\
-		Calcule la tensión de la cuerda con el método de multiplicadores de Lagrange.
-		La restricción es que la pesa se mantiene siempre en \(\vec{r} = \ell \hat{\rho}\), ergo la función que expresa esto es \(f(\rho) = \rho - \ell = 0\).
+		\textbf{Ideal pendulum}\\
+		Calculate the tension in the string using the method of Lagrange multipliers.
+		The constraint is that the bead is always at \(\vec{r} = \ell \hat{\rho}\), ergo, the function expressing this is \(f(\rho) = \rho - \ell = 0\).
 	\end{minipage}
 	\begin{minipage}[c][0cm][t]{0.25\textwidth}
 		\begin{tikzpicture}[scale= 1.0]
-  		\draw [arrows=-latex] (-1,2) -> (-1,1) node [above=15, right=2] {\(\vec{g}\)}; % g vertical
+  		\draw [arrows=-latex] (-1,2) -> (-1,1) node [above=15, right=2] {\(\vec{g}\)}; 
 			\draw [ultra thick] (-1.5,3) -- (2,3);
-			\fill [pattern = north east lines] (-1.5,3) rectangle (2,3.2); % techo
-			\draw [dashed] (0,3) -- (0,-.25);	% vertical
-			\draw [thick] (0,3) -- +(-60:3) node[midway,above,right=2] {\(\ell\)};	% inclinada +:relativa, -60 grados, longitud 3
+			\fill [pattern = north east lines] (-1.5,3) rectangle (2,3.2); 
+			\draw [dashed] (0,3) -- (0,-.25);
+			\draw [thick] (0,3) -- +(-60:3) node[midway,above,right=2] {\(\ell\)};	
 			\shade [ball color=black!80] ($(0,3)+(-60:3)$) circle(0.25) node [] {\color{white} $m$};
-    	\draw [arrows=-latex] (0,.4) -> (1.25,.4) node [midway, above] {\( \psi \)}; % desplazamiento horizontal
+    	\draw [arrows=-latex] (0,.4) -> (1.25,.4) node [midway, above] {\( \psi \)}; 
 			\draw [arrows=-latex] (0,0) arc [start angle=-90, end angle=-65, radius=3] node [below=12, left=8] {\( \varphi \)};
 		\end{tikzpicture}
 	\end{minipage}
@@ -33,12 +32,11 @@
 
 	\item 
 	\begin{minipage}[t][4cm]{0.47\textwidth}
-	\textbf{Cilindro que rueda por un plano inclinado} [Marion (e) ex. 7.5]\\
-	% English \textbf{Marion ejemplo 6.5 y ejemplo 7.9} Disco rodando en un plano inclinado.
+	\textbf{Cylinder rolling over an inclined plane} [Marion ex. 7.5]\\
 		\begin{enumerate}
-			\item Encuentre las ecuaciones de movimiento, 
-			\item la aceleración angular,
-			\item y la fuerzas de ligadura. 
[…]
