+++ translation.txt
@@ -57,8 +57,10 @@ fuerzasLigadura.ipynb               ->  forcesConstraint.ipynb
 guíaFuerzasLigadura.tex             ->  exercisesConstraintForces.tex   Ready.
 pénduloIdeal_FuerzasLigadura.ipynb  ->  constraintIdealPendulum.ipynb
 rodadura_FuerzasLigadura.ipynb      ->  rollingForcesConstraint.ipynb
-sistemasNoHolónomos.ipynb           ->  ?
+sistemasNoHolónomos.ipynb           ->  nonHolonomicSystems.ipynb
 
+
+https://stackoverflow.com/questions/35979642/what-is-git-tag-how-to-create-tags-how-to-checkout-git-remote-tags#35979751
 
 == 07NoConservativas  ->  07NonConservative
 
