+++ 04Constraints/exercisesConstraints.tex
@@ -36,9 +36,9 @@ Exercises marked with (*) have extra difficulty, don't hesitate to ask for help.
 	\item 
 	\begin{minipage}[t][2.3cm]{0.75\textwidth}
 		\textbf{Pendulum with sliding and coupled masses}\\ 
-		Two weights of masses \(m_1\) and \(m_2\) are linked together by a rigid bar of length \(\ell\) and negligible mass.
+		Two weights of masses \(m_1\) and \(m_2\) are linked together by a rigid rod of length \(\ell\) and negligible mass.
 		\(m_1\) can slide over an horizontal axis and \(m_2\) over a vertical axis, both without friction.
-		The bar sets a constraint between the coordinates that define their positions, \(x\) and \(y\).
+		The rod sets a constraint between the coordinates that define their positions, \(x\) and \(y\).
 	\end{minipage}
 	\begin{minipage}[c][1cm][t]{0.25\textwidth}
 		\input{figures/slidingWeights.tikz}
