+++ 08InertiaTensor/exercisesInertiaTensor.tex
@@ -1,6 +1,6 @@
 \documentclass[11pt, spanish, a4paper, twoside]{article}
 
-\input{../figuresLaTeX/introUNLaM}
+\input{../figurasLaTeX/introUNLaM}
 
 \begin{document}
 \begin{center}
@@ -12,11 +12,59 @@
 % Los problemas marcados con (*) son opcionales.
 
 \begin{enumerate}
+	\item
+	\textbf{Monóxido de carbono}\\
+	Calcular el tensor de inercia de una molécula requiere conocer la distancia entre los átomos y sus masas.
+
+	La distancia, o longitud de enlace, puede encontrarse en el \href{https://cccbdb.nist.gov/exp1x.asp}{buscador de propiedades de una molécula fruto de experimentos} de la \emph{Computational Chemistry Comparison and Benchmark DataBase} del National Institute of Standards and Technology, NIST, de los E.U.A.
+	Tras ingresar la fórmula química de la molécula, que en este caso es \emph{CO}, pueden encontrarse las distancias, o aún mejor las respectivas posiciones en un sistema de referecia cartesiano, expresadas en \si{angstrom}, equivalente a \SI{1e-10}{\metre}.
+
+	La masa para cada elemento químico está expresada en unidades de masa atómica unificada, \si{\atomicmassunit} en la \href{https://iupac.org/what-we-do/periodic-table-of-elements/}{tabla periódica} publicada por la Unión Internacional de Química Pura y Aplicada, IUPAC.
+	Esta es la masa expresada en gramos de un mol de átomos con la proporción de isótopos que se presenta en la naturaleza.
+	Para obtener la masa en gramos de uno sólo basta con dividirle por el número de átomos en este mol, la constante de Avogadro, \(N_A = \SI{6.022e23}{\per\mole}\).
+	
+	Expresar el tensor de inercia en unidades del SI (\si{\kilogram \squared \metre}).
+	Resultado:\\
+	\[
+		\overline{\overline{I}} = \left[\begin{matrix}0 & 0 & 0\\0 & 1.45 \cdot 10^{-46} & 0\\0 & 0 & 1.45 \cdot 10^{-46}\end{matrix}\right]
+	\]
+
+
 
 	\item 
-	\begin{minipage}[t][5.5cm]{0.65\textwidth}
+	\begin{minipage}[t][3cm]{0.73\textwidth}
+	\textbf{Agua}\\
+		Expresar el tensor de inercia en unidades del SI.
+		Resultado:\\
+		\[
+			\overline{\overline{I}} = \left[\begin{matrix}1.02 \cdot 10^{-47} & 0 & 0\\0 & 1.92 \cdot 10^{-47} & 0\\0 & 0 & 2.95 \cdot 10^{-47}\end{matrix}\right]
+		\]
+	\end{minipage}
+	\begin{minipage}[c][0.5cm][t]{0.2\textwidth}
+		\includegraphics[width=\textwidth]{moleculaH2O}
+	\end{minipage}
+
+	
+
+	\item 
+	\begin{minipage}[t][3cm]{0.73\textwidth}
+	\textbf{diclorometano}\\
+	Esta molécula tiene por fórmula química \isotope{CH_2Cl_2}. Expresar el tensor de inercia en unidades del SI.
+		Resultado:\\
+		\[
[…]
