+++ 08InertiaTensor/exercisesInertiaTensor.tex
@@ -1,11 +1,10 @@
-\documentclass[11pt, spanish, a4paper, twoside]{article}
+\documentclass[11pt, a4paper, twoside]{article}
 
-\input{../figurasLaTeX/introUNLaM}
+\input{../figuresLaTeX/introUNLaM}
 
 \begin{document}
 \begin{center}
-  % \textsc{\large Mecánica general}\\
-  \textsc{\large Cuerpo rígido | Tensores de inercia}
+  \textsc{\large Rigid body | Inertia tensors}
 \end{center}
 
 % De poder resolver estos problemas en forma autónoma puede asumir que adquirió los conocimientos mínimos sobre los temas abordados en la semana. No dude en consultar a docentes y compañeros si no puede terminarlos.
@@ -41,7 +40,7 @@
 		\]
 	\end{minipage}
 	\begin{minipage}[c][0.5cm][t]{0.2\textwidth}
-		\includegraphics[width=\textwidth]{moleculaH2O}
+		\includegraphics[width=\textwidth]{figures/moleculaH2O}
 	\end{minipage}
 
 	
@@ -56,7 +55,7 @@
 		\]
 	\end{minipage}
 	\begin{minipage}[c][0.5cm][t]{0.2\textwidth}
-		\includegraphics[width=\textwidth]{Dichloromethane_molecular_structure}
+		\includegraphics[width=\textwidth]{figures/Dichloromethane_molecular_structure}
 	\end{minipage}
 
 
@@ -74,7 +73,7 @@
 		\end{tasks}
 	\end{minipage}
 	\begin{minipage}[c][0cm][t]{0.3\textwidth}
-		\includegraphics[width=\textwidth]{o-021}
+		\includegraphics[width=\textwidth]{figures/o-021}
 	\end{minipage}
 	Resultados:\\
 			\[
