+++ 08InertiaTensor/exercisesInertiaTensor.tex
@@ -7,23 +7,18 @@
   \textsc{\large Rigid body | Inertia tensors}
 \end{center}
 
-% De poder resolver estos problemas en forma autónoma puede asumir que adquirió los conocimientos mínimos sobre los temas abordados en la semana. No dude en consultar a docentes y compañeros si no puede terminarlos.
-% Los problemas marcados con (*) son opcionales.
-
 \begin{enumerate}
 	\item
-	\textbf{Monóxido de carbono}\\
-	Calcular el tensor de inercia de una molécula requiere conocer la distancia entre los átomos y sus masas.
+	\textbf{Carbon monoxide}\\
+	Calculating the inertia tensor of a molecule requires knowing the distance between the atoms and their masses.
 
-	La distancia, o longitud de enlace, puede encontrarse en el \href{https://cccbdb.nist.gov/exp1x.asp}{buscador de propiedades de una molécula fruto de experimentos} de la \emph{Computational Chemistry Comparison and Benchmark DataBase} del National Institute of Standards and Technology, NIST, de los E.U.A.
-	Tras ingresar la fórmula química de la molécula, que en este caso es \emph{CO}, pueden encontrarse las distancias, o aún mejor las respectivas posiciones en un sistema de referecia cartesiano, expresadas en \si{angstrom}, equivalente a \SI{1e-10}{\metre}.
+	The distance, or bodn length, can be found at the \href{https://cccbdb.nist.gov/exp1x.asp}{Computational Chemistry Comparison and Benchmark DataBase} of the National Institute of Standards and Technology, NIST, U.S.A.
+	Querying about the chemical formula, \emph{CO} in this case, we can find the distances, or the positions in a cartesian frame of reference, expressed in \si{angstrom}, equivalent to \SI{1e-10}{\metre}.
 
-	La masa para cada elemento químico está expresada en unidades de masa atómica unificada, \si{\atomicmassunit} en la \href{https://iupac.org/what-we-do/periodic-table-of-elements/}{tabla periódica} publicada por la Unión Internacional de Química Pura y Aplicada, IUPAC.
-	Esta es la masa expresada en gramos de un mol de átomos con la proporción de isótopos que se presenta en la naturaleza.
-	Para obtener la masa en gramos de uno sólo basta con dividirle por el número de átomos en este mol, la constante de Avogadro, \(N_A = \SI{6.022e23}{\per\mole}\).
-	
-	Expresar el tensor de inercia en unidades del SI (\si{\kilogram \squared \metre}).
-	Resultado:\\
+	The mass for each chemical element is expressed in unified atomic mass units, \si{\atomicmassunit}, in the \href{https://iupac.org/what-we-do/periodic-table-of-elements/}{periodic table} published by the International Union of Pure and Applied Chemistry, IUPAC. This is the mass expressed in grams of one mole of atoms with the proportion of isotopes found in nature. To obtain the mass in grams of just one atom, simply divide by the number of atoms in this mole, Avogadro's constant, \(N_A = \SI{6.022e23}{\per\mole}\).
+
[…]
