### 2025-07-07 #256 @emckiernan
> [@emckiernan](https://github.com/emckiernan), I am also an editor of JOSE and have one JOSS review pending. If you can wait around a month for me to start, I could accept.

Hi, @nicoguaro. Circling back on this, a month has passed and we've been unable to find a second reviewer. Might you still be willing? Thanks in advance for considering!
