### 2025-07-10 #256 @emckiernan
Hi, @lauduvidovich, thank you for accepting my invitation to review over email! I'll tag you below to formally assign you. Here are the reviewer guidelines: https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html. I'll be closing this issue soon and opening up the review issue where you'll see a checklist. But if you have any questions, please just let me know. Thanks again!
