### 2025-07-10 #293 @editorialbot
[REVIEW]: Python-based Lagrange analytical mechanics course
**Submitting author:** @bettachini (<a href="http://orcid.org/0000-0001-7485-8884">Bettachini, V. A.</a>)
**Repository:** https://github.com/unlam/ComputationalAnalyticalMechanics/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
