### 2025-07-10 #293 @emckiernan
Hi, @richmf & @lauduvidovich, thanks again for agreeing to review! We'll be under this issue from now on. Please read over the reviewer guidelines https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html, and follow the instructions above under 'Checklists' to generate your personalized checklists. If you have questions, just let me know...I'm standing by to help!
