### 2025-07-10 ac5b963
Lanczos partial translation.
