+++ 03EulerLagrange/notesLanczos.tex
@@ -72,119 +72,115 @@
 \tableofcontents
 
 \section{Mechanics: vector vs. analytical approaches}
-\textbf{Referencia: Lanczos I\S1}\\
+\textbf{Reference: Lanczos I\S1}\\
 
 For Isaac Newton\footnote{English mathematician and physicist (Woolsthorpe, Lincolnshire, 1642 - Kensington, Middlesex 1726).}, the fundamental law for mechanics is the one that sets \(\va{F}_i= m_i \va{a}_i\) for particle \(i\).
 
 Thus, the linear momentum of this particle moving in free space is conserved until a force is exerted upon it. This approach is simple and adequate for a single particle.
 But when dealing with systems formed by many particles, we have to isolate each one of them and find the forces they exert upon each other. We call these forces \emph{interactions}.
 
-El desconocimiento de la naturaleza de muchas de estas fuerzas de interacción hacen necesario el uso de postulados adicionales.
-Newton pensó que su tercera ley, ``la acción equivale a la reacción'', se ocuparía de todos estos problemas dinámicos.
-Esto no resultó así, como demuestra el ejemplo de la dinámica de cuerpos rígidos para cuyo análisis hay que \emph{encorsetarse} en presunciones como de que las fuerzas entre componentes del sistema son todas centrales\footnote{En las fuerzas centrales estas apuntan de una partícula a la otra con la que interactúa, por tanto, la definición excluye a fuerzas observadas en el interior de sólidos, causadas por torsión o tensiones de corte que no son colineales con una recta pasando entre los elementos interactuantes.}.
+The lack of knowledge about the nature of many of these interaction forces makes the use of additional postulates necessary.
+Newton thought that his third law, ``action equals reaction'', would take care of all these dynamical problems.
+This did not turn out to be the case, as demonstrated by the example of rigid body dynamics, for whose analysis one must \emph{constrain oneself} to assumptions such as that the forces between components of the system are all central\footnote{In central forces these point from one particle to the other with which it interacts, therefore, the definition excludes forces observed inside solids, caused by torsion or shear stresses that are not collinear with a line passing between the interacting elements.}.
 
-En el enfoque analítico se analiza el \emph{sistema mecánico} en su conjunto, sin aislar fuerzas sobre partículas individuales.
-Si en el enfoque vectorial la fuerza sobre cada punto debe analizarse, en el analítico basta con conocer una única función que depende de la posición (y a veces también la velocidad) de las partículas; esta \emph{función trabajo} contiene en forma implícita todas las fuerzas que actúan sobre las partículas del sistema.
-Para obtener tales fuerzas basta una simple diferenciación de esta función.
+In the analytical approach, the \emph{mechanical system} is analyzed as a whole, without isolating forces on individual particles.
+If in the vector approach the force on each point must be analyzed, in the analytical approach it suffices to know a single function that depends on the position (and sometimes also the velocity) of the particles; this \emph{work function} contains in implicit form all the forces acting on the particles of the system.
+To obtain such forces, a simple differentiation of this function suffices.
 
-Otra ventaja de este último enfoque es el manejo de condiciones auxiliares.
-Muchas veces existen \emph{condiciones dinámicas} que se conocen \emph{a priori} del análisis de un sistema.
-Por ejemplo la distancia entre dos puntos cualesquiera de un cuerpo rígido se mantiene inalterable.
-Tal condición se mantiene gracias a fuertes fuerzas de vínculo entre partículas del sistema.
[…]
