### 2025-07-13 00e70b2
Translated using Claude, needs revision.
