+++ 03EulerLagrange/notesLanczos.tex
@@ -1,8 +1,6 @@
-\documentclass[12pt, spanish, a4paper, ]{article}
-% LTeX: language=es-AR
+\documentclass[12pt, english, a4paper]{article}
 
-\usepackage[spanish]{babel}
-\addto\shorthandsspanish{\spanishdeactivate{~<>}}
+\usepackage[english]{babel}
 \usepackage[utf8]{inputenc}
 
 \usepackage{float}
@@ -40,12 +38,11 @@
 \fancyhead{}
 \fancyfoot{\href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{$\vcenter{\hbox{\includegraphics[height=0.4cm]{by-nc-sa_80x15.pdf}}}$} \href{https://ingenieria.unlam.edu.ar/}{DIIT - UNLaM}}
 % \fancyfoot{{\textcopyright DIIT, UNLaM}}
-\fancyfoot[C]{ {\tiny Actualizado al \today} }
-\fancyfoot[R]{Pág. \thepage/\pageref{LastPage}}
+\fancyfoot[C]{ {\tiny Updated on \today} }
+\fancyfoot[R]{Page \thepage/\pageref{LastPage}}
 \renewcommand{\headrulewidth}{0pt}
 \renewcommand{\footrulewidth}{0pt}
 
-% \usepackage{multicol} % lista multicolumna
 
 \usepackage{changes} % textsub super script
 
@@ -469,23 +466,14 @@ This reformulation of d'Alembert's principle called \emph{Hamilton's principle} 
 This procedure, of finding the equations that describe the system's dynamics through a minimization of the action, or \emph{principle of least action}, is common to classical, relativistic, and wave mechanics.
 The only thing that changes in each approach is how the Lagrangian is written.
 
+\section{Euler-Lagrange Equation}\label{EulerLagrange} % Lanczos II.10 % Landau \SI.2
+\textbf{Reference: Lanczos II\S10, Landau \S I.2}\\
 
+How to obtain the stationary value of the action \(S\)?
+The answer to how to ensure the condition \(\delta S= 0\) is linked to the concept of virtual displacements (see section \ref{desplazamientoVirtual}).
+It is precisely about exploring the set of variations over \((q_i, \dot{q}_i, t)\) that effectively ensure minimization of \(S= \int \mathcal{L}(q_i, \dot{q}_i, t) \dd t\).
 
-\section{Ecuación de Euler-Lagrange}\label{EulerLagrange} % Lanczos II.10 % Landau \SI.2
-\textbf{Referencia: Lanczos II\S10, Landau \S I.2}\\
-  
-
-
-%De producir un pequeño despalazmiento en \(q\) hasta \(q+ \delta q\) la variación en la acción \(S\) es
-%\begin{equation}
-%    L= L(q_i, \dot{q}_i, t) = T - V,
-%    \label{Landau11}
-%\end{equation}
-¿Cómo obtener el valor estacionario de la acción \(S\)?
-La respuesta a como asegurar la condición \(\delta S= 0\) está ligada al concepto de desplazamientos virtuales (ver sección \ref{desplazamientoVirtual}).
-Se trata justamente de explorar el conjunto de variaciones sobre \((q_i, \dot{q}_i, t)\) que efectivamente aseguren minimizar \(S= \int \mathcal{L}(q_i, \dot{q}_i, t) \dd t\).
-
-Empezamos analizando la variación del integrando de \(S\), es decir el Lagrangiano ante una variación de coordenadas generalizadas, \(\overline{q_i (t)}= q_i(t) + \delta q_i = q_i (t)+ \epsilon \tilde{q}_i (t)\), y velocidades generalizadas
[…]
