### 2025-07-18 #293 @emckiernan
Hi, @richmf & @lauduvidovich. Just checking in to see if you have any questions?
