### 2025-07-23 d5e03aa
Ready, menos volanteInercia.
