+++ 10EulerRotation/exercisesEulerRotation.tex
@@ -1,41 +1,55 @@
-\documentclass[11pt, spanish, a4paper, twoside]{article}
+\documentclass[11pt, english, a4paper, twoside]{article}
 
 \input{../figuresLaTeX/introUNLaM}
 
 \begin{document}
 \begin{center}
-  % \textsc{\large Mecánica general}\\
-  \textsc{\large Cuerpo rígido | Ecuaciones de Euler}
+  \textsc{\large Rigid Body | Euler's Equations}
 \end{center}
 
-% De poder resolver estos problemas en forma autónoma puede asumir que adquirió los conocimientos mínimos sobre los temas abordados en la semana. No dude en consultar a docentes y compañeros si no puede terminarlos. Los problemas marcados con * son opcionales.
 
 \begin{enumerate}
 
 	\item 
 	\begin{minipage}[t][5cm]{0.55\textwidth}
-		\textbf{Engranaje inclinado}
-		Un engranaje de masa de \SI{10}{\kilo\gram} está montado con una inclinación de \ang{10;;} en un eje de masa despreciable.
-		Los cojinetes \(A\) y \(B\) sostienen el eje que gira con velocidad angular constante.
-		El \(A\) es de empuje, por lo que provee reacción también en la dirección longitudinal al eje en tanto que el \(B\) solo lo hace en las direcciones transversales.
-		Los momentos de inercia del engranaje son \(I_z = \SI{.1}{\kilo\gram\metre\squared}\) y el \(I_y = \SI{.05}{\kilo\gram\metre\squared}\).
+		\textbf{Inclined Gear}
+		A gear with a mass of \SI{10}{\kilo\gram} is mounted with an inclination of \ang{10;;} on a shaft of negligible mass.
+		Bearings \(A\) and \(B\) support the shaft which rotates at constant angular velocity.
+		Bearing \(A\) is a thrust bearing, so it provides reaction also in the longitudinal direction of the shaft, while bearing \(B\) only does so in the transverse directions.
+		The moments of inertia of the gear are \(I_z = \SI{.1}{\kilo\gram\metre\squared}\) and \(I_y = \SI{.05}{\kilo\gram\metre\squared}\).
 		\begin{tasks} 
-			\task Determine las reacciones que deben proveer los cojinetes para el instante en que el sistema en rotación presenta la disposición que se ilustra.
+			\task Determine the reactions that the bearings must provide for the instant when the rotating system presents the arrangement shown.
 		\end{tasks}
 	\end{minipage}
 	\begin{minipage}[c][0.5cm][t]{0.4\textwidth}
-		\includegraphics[width=\textwidth]{figures/hibb21_4}
+		\includegraphics[width=\textwidth]{figures/hibb_21-4}
 	\end{minipage}
 
 
+	\item 
+	\begin{minipage}[t][4.6cm]{0.55\textwidth}
+		\textbf{Flywheel}
+		The flywheel centered at \(G\) has a mass of \SI{10}{\kilo\gram} and is integral with the shaft of negligible mass that rotates at constant angular velocity \(\omega_s= \SI{6}{\per\second} \) (radians per second) supported by bearings \(A\) and \(B\).
+		The first is a thrust bearing, so it provides reaction also in the longitudinal direction of the shaft, while the second only does so in the transverse directions.
+		A transverse shaft to that of the flywheel supports the mount of bearing \(A\) and also rotates at constant angular velocity \(\omega_p\).
+		\begin{tasks}
[…]
