+++ 12OscillationsMDOF/exercisesMDOF.tex
@@ -1,6 +1,7 @@
 \documentclass[11pt, spanish, a4paper, twopage]{article}
+% LTeX: language=es-AR
 
-\input{../figuresLaTeX/introUNLaM}
+\input{../figurasLaTeX/introUNLaM}
 
 \begin{document}
 \begin{center}
@@ -15,29 +16,30 @@
 				
 \item 
 \begin{minipage}[t][4cm]{0.65\textwidth}
-En el sistema de la figura \(l = \SI{0.5}{\metre}\), \(k_1 = k_2 = k = \SI{2E3}{\newton\per\metre}\) y \(m_1 = m_2 = m_3 = m = \SI{1}{\kilo\gram}\).
-Asumiendo pequeñas las oscilaciones en torno al cero de las coordenadas indicadas:
-\begin{tasks}
-	\task obtenga las ecuaciones de Euler-Lagrange,
-	\task escríbalas en forma matricial (matrices M, K), y
-	\task obtenga las frecuencias naturales de oscilación del sistema.
-\end{tasks}
+	\textbf{Múltiples carros}
+	En el sistema de la figura \(l = \SI{0.5}{\metre}\), \(k_1 = k_2 = k = \SI{2E3}{\newton\per\metre}\) y \(m_1 = m_2 = m_3 = m = \SI{1}{\kilo\gram}\).
+	Asumiendo pequeñas las oscilaciones en torno al cero de las coordenadas indicadas:
+	\begin{tasks}
+		\task obtenga las ecuaciones de Euler-Lagrange,
+		\task escríbalas en forma matricial (matrices M, K), y
+		\task obtenga las frecuencias naturales de oscilación del sistema.
+	\end{tasks}
 \end{minipage}
 \begin{minipage}[c][0cm][t]{0.3\textwidth}
-	\includegraphics[width=\textwidth]{figures/shabana_fig_P3_6.png}
+	\includegraphics[width=\textwidth]{shabana_fig_P3_6.png}
 \end{minipage}
 
 
 
 \item 
 \begin{minipage}[t][2.5cm]{0.65\textwidth}
-\textbf{Péndulo de torsión compuesto}
+	\textbf{Péndulo de torsión compuesto}
 
-El sistema de la figura consiste en un eje atraviesa tres discos que tienen por momentos de inercia \(I_1, I_2\) e \(I_3\) todos de igual magnitud \SI{1E3}{\kilo\gram\metre\squared}.
-El eje de acero tiene un diámetro \(d = \SI{0.01}{\metre}\) y sus secciones longitudes de \(l_1 = l_2 = l_3 = \SI{0.5}{\metre}\).
+	El sistema de la figura consiste en un eje atraviesa tres discos que tienen por momentos de inercia \(I_1, I_2\) e \(I_3\) todos de igual magnitud \SI{1E3}{\kilo\gram\metre\squared}.
+	El eje de acero tiene un diámetro \(d = \SI{0.01}{\metre}\) y sus secciones longitudes de \(l_1 = l_2 = l_3 = \SI{0.5}{\metre}\).
 \end{minipage}
 \begin{minipage}[c][1cm][t]{0.3\textwidth}
-	\includegraphics[width=\textwidth]{figures/shabana_fig_P3_5.png}
+	\includegraphics[width=\textwidth]{shabana_fig_P3_5.png}
 \end{minipage}
 
 
