### 2025-07-23 cafde09
12 translating
