+++ 12OscillationsMDOF/exercisesMDOF.tex
@@ -1,83 +1,64 @@
-\documentclass[11pt, spanish, a4paper, twopage]{article}
-% LTeX: language=es-AR
-
-\input{../figurasLaTeX/introUNLaM}
-
+\documentclass[11pt, english, a4paper, twopage]{article}
+\input{../figuresLaTeX/introUNLaM}
 \begin{document}
 \begin{center}
-	% \textsc{\large Mecánica general}\\
-	\textsc{\large Vibraciones en sistemas discretos | Múltiples grados de libertad}
+ \textsc{\large Oscillations in discrete systems | Multiple degrees of freedom}
 \end{center}
-
-% De poder resolver estos problemas en forma autónoma puede asumir que adquirió los conocimientos mínimos sobre los temas abordados en la semana. No dude en consultar a docentes y compañeros si no puede terminarlos. Los problemas marcados con * son opcionales.
-
 \begin{enumerate}
-
-				
-\item 
+\item
 \begin{minipage}[t][4cm]{0.65\textwidth}
-	\textbf{Múltiples carros}
-	En el sistema de la figura \(l = \SI{0.5}{\metre}\), \(k_1 = k_2 = k = \SI{2E3}{\newton\per\metre}\) y \(m_1 = m_2 = m_3 = m = \SI{1}{\kilo\gram}\).
-	Asumiendo pequeñas las oscilaciones en torno al cero de las coordenadas indicadas:
-	\begin{tasks}
-		\task obtenga las ecuaciones de Euler-Lagrange,
-		\task escríbalas en forma matricial (matrices M, K), y
-		\task obtenga las frecuencias naturales de oscilación del sistema.
-	\end{tasks}
+ \textbf{Multiple carts}
+ In the system shown in the figure \(l = \SI{0.5}{\metre}\), \(k_1 = k_2 = k = \SI{2E3}{\newton\per\metre}\) and \(m_1 = m_2 = m_3 = m = \SI{1}{\kilo\gram}\).
+ Assuming small oscillations around zero of the indicated coordinates:
+ \begin{tasks}
+ \task obtain the Euler-Lagrange equations,
+ \task write them in matrix form (matrices M, K), and
+ \task obtain the natural oscillation frequencies of the system.
+ \end{tasks}
 \end{minipage}
 \begin{minipage}[c][0cm][t]{0.3\textwidth}
-	\includegraphics[width=\textwidth]{shabana_fig_P3_6.png}
+ \includegraphics[width=\textwidth]{figures/shabana_fig_P3_6.png}
 \end{minipage}
-
-
-
-\item 
+\item
 \begin{minipage}[t][2.5cm]{0.65\textwidth}
-	\textbf{Péndulo de torsión compuesto}
-
-	El sistema de la figura consiste en un eje atraviesa tres discos que tienen por momentos de inercia \(I_1, I_2\) e \(I_3\) todos de igual magnitud \SI{1E3}{\kilo\gram\metre\squared}.
-	El eje de acero tiene un diámetro \(d = \SI{0.01}{\metre}\) y sus secciones longitudes de \(l_1 = l_2 = l_3 = \SI{0.5}{\metre}\).
+ \textbf{Compound torsional pendulum}
+ The system in the figure consists of a shaft that passes through three disks having moments of inertia \(I_1, I_2\) and \(I_3\) all of equal magnitude \SI{1E3}{\kilo\gram\metre\squared}.
+ The steel shaft has a diameter \(d = \SI{0.01}{\metre}\) and its sections have lengths of \(l_1 = l_2 = l_3 = \SI{0.5}{\metre}\).
 \end{minipage}
[…]
