### 2025-07-23 32afa25
12 Ready
