+++ translation.txt
@@ -14,7 +14,12 @@ TODO:
   After 8 or 9, it's translated from current version.
 - Translate imagesNotebooks.ipynb
 
-- Español: cambiar "o" con \circ.
+
+- Check figures: change links to embedded.
+- Test everything in Colab.
+
+- Incluir NoInercial cuando esté listo.
+
 
 
 
@@ -28,7 +33,7 @@ guíaVectorial.tex          ->  exercisesVector.tex      Ready.
 
 == 02Energía -> 02Energy
 
-apunteLanczos.pdf                  ->  notesLanczos.pdf                     Ver .tex en 03
+apunteLanczos.pdf                  ->  notesLanczos.pdf                     Ready
 energíaCinéticaTraslación.ipynb    ->  kineticEnergy.ipynb                  Ready.
 energíaPotencialGravitatoria.ipynb ->  gravitationalPotentialEnergy.ipynb   Ready.
 guíaEnergía.tex                    ->  exercisesEnergy.tex                  Ready.
@@ -36,7 +41,7 @@ guíaEnergía.tex                    ->  exercisesEnergy.tex                  Re
 
 == 03EulerLagrange -> 03EulerLagrange
 
-apunteLanczos.tex                              ->  notesLanczos.tex               Translated using Claude, to be checked.
+apunteLanczos.tex                              ->  notesLanczos.tex               Ready - Translated using Claude, to be checked.
 energíaPotencialElástica_eulerLagrange.ipynb   ->  elasticPotentialEnergy.ipynb   Ready.
 guíaEulerLagrange.tex                          ->  exercisesEulerLagrange.tex     Ready (suggestion 4, ref.)
 partesRotantes_eulerLagrange.ipynb             ->  rotatingParts.ipynb            Ready
@@ -78,13 +83,13 @@ noConservativas.ipynb                            ->  nonconservative.ipynb      
 
 guíaTensorInercia.tex               -> exercisesInertiaTensor.tex           Ready
 momentoAngularTorque.ipynb          -> angular_momentum_torque.ipynb        Ready
-monóxidoCarbono_tensorInercia.ipynb -> carbonMonoxide_inertiaTensor.ipynb   Almost ready (18)
+monóxidoCarbono_tensorInercia.ipynb -> carbonMonoxide_inertiaTensor.ipynb   Ready
 tensorInercia.ipynb                 -> inertiaTensor.ipynb                  (19)
 
 
 == 09MasaDistribuida  -> 09DistributedMass (21)
 
-guíaDistribuciónMasa.tex   ->  exercisesDistributedMass.tex   Almost ready (20)
+guíaDistribuciónMasa.tex   ->  exercisesDistributedMass.tex   Ready
 cubo_tensorInercia.ipynb   ->  inertiaTensorcube.ipynb        Ready
 masaDistribuida.ipynb      ->  distributedMass.ipynb          Ready
 
@@ -94,7 +99,7 @@ masaDistribuida.ipynb      ->  distributedMass.ipynb          Ready
 engranajeDesalineado.ipynb  ->  misalignedGear.ipynb          Ready
 guíaRotacionEuler.tex       ->  exercisesEulerRotation.tex    Ready    
 rotaciónEuler.ipynb         ->  rotationEuler.ipynb           Ready
-volanteInercia.ipynb        ->  flywheel.ipynb                (21) ¿Borrar?
+volanteInercia.ipynb        ->  flywheel.ipynb                Ready (21)
 
 
 ==  11Vibraciones1GdL  ->  11OscillationsSDOF
@@ -106,9 +111,9 @@ vibraciones1GdL.ipynb    ->  oscillationsSDOF.ipynb       Ready
 
[…]
