### 2025-07-24 cff46f5
Fixed: figures, typos, spanish words, inertiaTensor.
