+++ 05Simulation/exercisesSimulation.tex
@@ -5,7 +5,6 @@
 
 \begin{document}
 \begin{center}
-  % \textsc{\large Mecánica general}\\
   \textsc{\large Simulation | Numerical solutions for the Euler-Lagrange equation}
 \end{center}
 
