### 2025-07-24 ce9db48
Replaced link to license in every ipynb.
