+++ 11OscillationsSDOF/exercisesSDOF.tex
@@ -23,7 +23,7 @@
  Use the quarter car model, in which it is assumed that each shock absorber supports one quarter of the weight.
  Further simplify this model by eliminating the tire, both its mass and its ability to operate as a shock absorber, to assign this latter task solely to the suspension.
  Since driving in the overdamped regime is uncomfortable, as after a pothole a violent rebound can occur, you must adjust the suspension accordingly.
- From a \href{https://journals.sagepub.com/doi/pdf/10.1177/1687814016648638}{paper in \emph{Advances in Mechanical Engineering}} we take a standard value of $k_s = \SI{12500}{\newton\per\metre}$ for the shock absorber
+ From a \href{https://journals.sagepub.com/doi/pdf/10.1177/1687814016648638}{paper in \emph{Advances in Mechanical Engineering}} we take a standard value of $k_s = \SI{12500}{\newton\per\metre}$ for the shock absorber.
 \end{minipage}
 \begin{minipage}[c][0cm][t]{0.3\textwidth}
  \includegraphics[width=\textwidth]{figures/amortiguadores_AudiCoupeTT}
