+++ figuresLaTeX/introUNLaM.tex
@@ -50,7 +50,7 @@
 \fancyhead[LE, LO]{\textbf{Computational Analytical Mechanics} }
 % \fancyhead[LE, LO]{\textbf{Mecánica General} }
 \fancyhead[RE, RO]{\href{https://ingenieria.unlam.edu.ar/}{$\vcenter{\hbox{\includegraphics[height=1cm]{ambos.pdf}}}$}}
-\fancyfoot{\href{https://creativecommons.org/licenses/by-nc-sa/4.0/deed.es_ES}{$\vcenter{\hbox{\includegraphics[height=0.4cm]{by-nc-sa_80x15.pdf}}}$} \href{https://ingenieria.unlam.edu.ar/}{DIIT - UNLaM}}
+\fancyfoot{\href{https://creativecommons.org/licenses/by-nc-sa/4.0/}{$\vcenter{\hbox{\includegraphics[height=0.4cm]{by-nc-sa_80x15.pdf}}}$} \href{https://ingenieria.unlam.edu.ar/}{DIIT - UNLaM}}
 \fancyfoot[C]{ {\tiny Updated \today} }
 \fancyfoot[RO, LE]{Page \thepage/\pageref{LastPage}}
 \renewcommand{\headrulewidth}{0pt}
