+++ translation.txt
@@ -3,8 +3,7 @@ tags/202503
 
 TODO:
 - Informatica: ¿qué hacer? ¿Se traduce o no se incluye?
-
-- Check Link to license in Spanish, some are not working.
+- Check the notes for Readme below.
 
 - Test everything in Colab.
 
@@ -12,6 +11,21 @@ TODO:
 - Edit tensorInercia.ipynb accordingly (see 19 below) 
 
 
+===================================
+Readme
+
+
+The references to IntroPython and Markdown are commented out in the Readme file, waiting for a decision about the folder Informatica.
+
+Added link to 06 nonHolonomicSystems.ipynb
+
+Commented out 97 nonConservativesExamples.ipynb
+
+Commented out 08 ex3RotationArm.ipynb and ex5Cube.ipynb (examples).
+
+Added link to 08 carbonMonoxide_inertiaTensor.ipynb
+
+Added link to 09 distributedMass.ipynb and inertiaTensorcube.ipynb
 
 
 
