+++ docs/index.md
@@ -1,18 +1,21 @@
+<!-- Remove for github pages -->
+<!-- [<image src="docs/assets/img/bothLogo.svg" alt="Logo UNLAM"  width="200"/>](https://ingenieria.unlam.edu.ar/) -->
+
+<!-- Uncomment for github pages -->
 ![https://ingenieria.unlam.edu.ar/](/docs/assets/img/bothLogo.svg )
-<!--
-[<image src="{{site.url}}/docs/assets/img/bothLogo.svg" alt="Logo UNLAM"  width="200"/>](https://ingenieria.unlam.edu.ar/)
--->
+
 
 # Computational Analytical Mechanics
 
+<!-- Remove for github pages -->
+<!-- [<image src="https://upload.wikimedia.org/wikipedia/commons/1/12/Cc-by-nc-sa_icon.svg"  width="80"/>](https://creativecommons.org/licenses/by-nc-sa/4.0/) 2024 Víctor A. Bettachini -->
 
+<!-- Uncomment for github pages -->
 ![https://creativecommons.org/licenses/by-nc-sa/4.0/](/docs/assets/img/by-nc-sa.svg) 2024 Víctor A. Bettachini
 
-<!--
-<img src="/docs/assets/img/by-nc-sa.src"/>
+<!-- Remove for github pages -->
+<!-- Access to the material: [https://unlam.github.io/ComputationalAnalyticalMechanics/](https://unlam.github.io/ComputationalAnalyticalMechanics/) -->
 
-[<image src="https://upload.wikimedia.org/wikipedia/commons/1/12/Cc-by-nc-sa_icon.svg"  width="80"/>](https://creativecommons.org/licenses/by-nc-sa/4.0/) 2024 Víctor A. Bettachini
--->
 
 This is the repository for the course [Computational Analytical Mechanics](https://ingenieria.unlam.edu.ar/preview.php?seccion=3&idArticulo=510), for the [mechanical engineering degree](https://ingenieria.unlam.edu.ar/index.php?seccion=3&idArticulo=371) at the [Department of Engineering and Research in Technology](https://ingenieria.unlam.edu.ar/), Universidad Nacional de La Matanza.
 
@@ -40,7 +43,7 @@ Some of them are covered through more than one week, so a further descriptor, _F
 | 10   | "                     | 10EulerRotation    | Euler equations for the rigid body. |
 | 11   | "                     | "                | **Final project: statement of the problem** |
 | 12   | Oscillations          | 11OscillationsSDOF | Forced oscillations in single degree of freedom systems. |
-| 13   | "                     | 12OscillationsMDOF | Forced oscillations at multiple degrees of freedom systems. |
+| 13   | "                     | 12OscillationsMDOF | Forced oscillations in multiple degrees of freedom systems. |
 | 14   | "                     | "                | "                                                           |
 | 15   | Evaluation            | (none)           | **Project live presentation**                                                   |
 | 16   | "                     | (none)           | **2.nd chance presentation**                                                   |
@@ -81,9 +84,9 @@ Some of them are covered through more than one week, so a further descriptor, _F
 - Exercises
   - [Problem set - Vector kinematics](https://github.com/unlam/ComputationalAnalyticalMechanics/blob/main/01Vector/exercisesVector.pdf) [pdf]
 
-- References
+<!-- - References
   - [Introduction to Python](https://colab.research.google.com/github/unlam/ComputationalAnalyticalMechanics/blob/main/01Vector/IntroPython.ipynb) [ipynb]
[…]
