### 2025-07-24 a0a0ba0
Fixing logo in page.
