### 2025-07-24 036aaf5
Fixing license logo in webpage.
