### 2025-07-24 e5cd57d
Fixed logo in webpage.
