### 2025-07-24 af5caa2
Fixed spelling: nonconservative. Removed translating.txt and changed it to todo.txt.
