+++ translation.txt
@@ -1,212 +1,0 @@
-
-tags/202503
-
-TODO:
-- Informatica: ¿qué hacer? ¿Se traduce o no se incluye?
-- Check the notes for Readme below.
-
-- Test everything in Colab.
-
-- Include NonInertial when ready.
-- Edit tensorInercia.ipynb accordingly (see 19 below) 
-
-
-===================================
-Readme
-
-
-The references to IntroPython and Markdown are commented out in the Readme file, waiting for a decision about the folder Informatica.
-
-Added link to 06 nonHolonomicSystems.ipynb
-
-Commented out 97 nonConservativesExamples.ipynb
-
-Commented out 08 ex3RotationArm.ipynb and ex5Cube.ipynb (examples).
-
-Added link to 08 carbonMonoxide_inertiaTensor.ipynb
-
-Added link to 09 distributedMass.ipynb and inertiaTensorcube.ipynb
-
-
-
-===================================================================
-
-== 01Vectorial -> 01Vector 
-
-cinemáticaVectorial.ipynb  ->  vectorKinematics.ipynb   Ready
-estaAsignatura.ipynb       ->  thisCourse.ipynb         Ready
-guíaVectorial.tex          ->  exercisesVector.tex      Ready.
-
-
-== 02Energía -> 02Energy
-
-apunteLanczos.pdf                  ->  notesLanczos.pdf                     Ready
-energíaCinéticaTraslación.ipynb    ->  kineticEnergy.ipynb                  Ready.
-energíaPotencialGravitatoria.ipynb ->  gravitationalPotentialEnergy.ipynb   Ready.
-guíaEnergía.tex                    ->  exercisesEnergy.tex                  Ready.
-
-
-== 03EulerLagrange -> 03EulerLagrange
-
-apunteLanczos.tex                              ->  notesLanczos.tex               Ready - Translated using Claude, to be checked.
-energíaPotencialElástica_eulerLagrange.ipynb   ->  elasticPotentialEnergy.ipynb   Ready.
-guíaEulerLagrange.tex                          ->  exercisesEulerLagrange.tex     Ready (suggestion 4, ref.)
-partesRotantes_eulerLagrange.ipynb             ->  rotatingParts.ipynb            Ready
-plantilla_eulerLagrange.ipynb                  ->  euler_lagrange_base.ipynb      Ready
-péndulos_eulerLagrange.ipynb                   ->  eulerLagrangePendulums.ipynb   Ready
-
-
-== 04Ligaduras  ->  04Constraints
-
-atwood_Ligaduras.ipynb            ->  atwoodConstraint.ipynb        Ready (suggestion Test that everything works)
-guíaLigaduras.tex                 ->  exercisesConstraints.tex      Ready (suggestion 6)
-ligadurasFunción.ipynb            ->  constraintsFunction.ipynb     Ready (suggestion 8)
-resoluciónSistemasLineales.ipynb  ->  solvingLinearSystems.ipynb    Ready
-figures\ringPulley.tikz: changed labels.
-
-
-== 05Simulación  ->  05Simulation
-
-atwoodSimulación_Resuelto.ipynb         ->  atwoodSimulationSolved.ipynb   Ready
-guíaSimulación.tex                      ->  exercisesSimulation.tex        Ready
-pénduloLibre_Simulación_Resuelto.ipynb  ->  pendulumFreeSimulSolved.ipynb  Ready (comment 12)
-
-== 06FuerzasLigadura  ->  06ForcesConstraint
-
-fuerzasLigadura.ipynb               ->  forcesConstraint.ipynb          Ready (comment check (14) (from 2021))
[…]
