### 2025-07-24 c6557ee
Changed expression: misaligned gear.
