+++ todo.txt
@@ -1,9 +1,12 @@
 
 TODO:
 - Check the notes for Readme below.
-
+  - I (VB) do not know what to check.
 - Test everything in Colab.
-
+  - 01Vector | done
+  - 02Energy | done
+  - 03EulerLagrange | done
+  - 04Constraints | done
 - Include NonInertial when ready.
 - Edit tensorInercia.ipynb accordingly (see 19 below) 
 
