### 2025-07-28 f6f0195
nonconservative.ipynb colab array rendering
