+++ todo.txt
@@ -7,6 +7,9 @@ TODO:
   - 02Energy | done
   - 03EulerLagrange | done
   - 04Constraints | done
+  - 05Simulation | done
+  - 06ForcesConstraint | done
+  
 - Include NonInertial when ready.
 - Edit tensorInercia.ipynb accordingly (see 19 below) 
 
