### 2025-07-28 a725d1e
todo | checked all ipynb run in Colab
