### 2025-09-10 #293 @emckiernan
Hi, @richmf & @lauduvidovich. My apologies to keep bugging you both, but please submit your reviews as soon as possible. Today marks two months since we opened the review, and I'd really love to get the authors some feedback. @richmf, I see you opened your checklist, please let me know if you have questions on any of the items. Thanks in advance!
