### 2025-09-11 06ef3fc
04Constraints Ring and pulley result expression
