+++ 04Constraints/exercisesConstraints.tex
@@ -71,7 +71,8 @@ Exercises marked with (*) have extra difficulty, don't hesitate to ask for help.
 		\item Write the Euler-Lagrange equation.\\
 		Result:
 		\(
-		R_{r}^{2} m_1 \ddot{\theta} + R_{r}^{2} m_{r} \ddot{\theta} + R_{r} g m_1 \cos{\left(\theta \right)} + R_{p}^{2} m_{2} \ddot{\theta} + \frac{R_{p}^{2} m_{p} \ddot{\theta}}{2} - R_{p} g m_{2} = 0
+		- R_{r}^{2} m_1 \ddot{\theta} - R_{r}^{2} m_{r} \ddot{\theta} - R_{r}^{2} m_2 \ddot{\theta} - \frac{R_{r}^{2} m_{p} \ddot{\theta}}{2} + R_{r} g m_2 \cos{\left(\theta \right)} - R_{r} g m_2 = 0
+		% R_{r}^{2} m_1 \ddot{\theta} + R_{r}^{2} m_{r} \ddot{\theta} + R_{r} g m_1 \cos{\left(\theta \right)} + R_{p}^{2} m_{2} \ddot{\theta} + \frac{R_{p}^{2} m_{p} \ddot{\theta}}{2} - R_{p} g m_{2} = 0
 		\)
 	\end{enumerate}
 
