### 2025-09-12 #293 @richmf
Hi
I'm working on it. I'm working on it. Sorry for the delay, the review will
be ready on Sunday.
Saludos cordiales,

Dr. Ricardo Méndez-Fragoso

Cubículo 411. Departamento de Física

Facultad de Ciencias, UNAM.

Office 411. Physics Department.

School of Sciences, UNAM.

Telegram: @elrichmf

Web: https://sistemas.fciencias.unam.mx/rich
<http://sistemas.fciencias.unam.mx/rich>

On Wed, Sep 10, 2025 at 4:17 PM Erin C. McKiernan ***@***.***>
wrote:

> *emckiernan* left a comment (openjournals/jose-reviews#293)
> <https://github.com/openjournals/jose-reviews/issues/293#issuecomment-3276715763>
>
> Hi, @richmf <https://github.com/richmf> & @lauduvidovich
> <https://github.com/lauduvidovich>. My apologies to keep bugging you
> both, but please submit your reviews as soon as possible. Today marks two
> months since we opened the review, and I'd really love to get the authors
> some feedback. @richmf <https://github.com/richmf>, I see you opened your
> checklist, please let me know if you have questions on any of the items.
> Thanks in advance!
>
