### 2025-09-16 #293 @richmf
Hi @emckiernan

I've finished the review. I'm so sorry for the delay, but this is a new way of doing these reviews, and I also had a heavy workload.

I'm happy to say I've already finished the review. As feedback, I only offer the following recommendations to the author:
- The README file should include all the dependencies needed for the entire repository. For example:
"To use this repository, you need at least the following libraries:
- numpy 2.3.0
- scipy 1.16.1
etc.
"
- The README file should include a space for community feedback, either by email or through GitHub messages.
- Not all references have DOIs, and they aren't included in the .md file.
- In the JOSE paper (PDF) there is a phrase in spanish in the bibliography section that not make sense
I consider that the corrections are minor and can be given the green light for acceptance.
 Best regards,
