### 2025-09-19 #293 @emckiernan
> Hi, [@ntmoore](https://github.com/ntmoore). I'm reaching out because you previously signed up to review for JOSE. We're looking for another reviewer for this submission. Would you be willing to contribute a review? Please let me know if you're willing and available. Thanks in advance for considering!

Hi @ntmoore, just following up to see if you might be willing to review this, or have a recommendation for another reviewer? Thanks in advance!
