### 2025-09-22 #293 @bettachini
> [@richmf](https://github.com/richmf), thanks so much! [@bettachini](https://github.com/bettachini), please let us know if you have any questions about the above comments. In the meantime, I'll continue to work on finding us another reviewer. Thanks!

No major questions besides one which answer is implicit, that I should proceed at once to perform the recommended modifications, to both the repository and the paper.
