### 2025-10-15 #293 @emckiernan
Hi, @hughshanahan. I'm reaching out because you previously signed up to review for JOSE. We're looking for another reviewer for this submission. Would you be willing to contribute a review? Please let me know if you're willing and available. Thanks in advance for considering!
