### 2025-10-22 24dc95f
paper/paper.md
