+++ README.md
@@ -14,8 +14,9 @@
 <!-- 2024 Víctor A. Bettachini -->
 
 <!-- Remove for github pages -->
+<!--
 Access to the material: [https://unlam.github.io/ComputationalAnalyticalMechanics/](https://unlam.github.io/ComputationalAnalyticalMechanics/)
-
+--->
 
 This is the repository for the course [Computational Analytical Mechanics](https://ingenieria.unlam.edu.ar/preview.php?seccion=3&idArticulo=510), for the [mechanical engineering degree](https://ingenieria.unlam.edu.ar/index.php?seccion=3&idArticulo=371) at the [Department of Engineering and Research in Technology](https://ingenieria.unlam.edu.ar/), Universidad Nacional de La Matanza.
 A [Spanish version](https://github.com/unlam/MecanicaAnaliticaComputacional) is available.
