+++ docs/index.md
@@ -1,241 +1,0 @@
-<!-- Remove for github pages -->
-<!-- [<image src="docs/assets/img/bothLogo.svg" alt="Logo UNLAM"  width="200"/>](https://ingenieria.unlam.edu.ar/) -->
-
-<!-- Uncomment for github pages -->
-[<image src="assets/img/bothLogo.svg" alt="Logo UNLAM"  width="200"/>](https://ingenieria.unlam.edu.ar/)
-
-# Computational Analytical Mechanics
-
-<!-- Remove for github pages -->
-<!-- [<image src="https://upload.wikimedia.org/wikipedia/commons/1/12/Cc-by-nc-sa_icon.svg"  width="80"/>](https://creativecommons.org/licenses/by-nc-sa/4.0/) 2024 Víctor A. Bettachini -->
-
-<!-- Uncomment for github pages -->
-[<image src="https://upload.wikimedia.org/wikipedia/commons/1/12/Cc-by-nc-sa_icon.svg"  width="80"/>](https://creativecommons.org/licenses/by-nc-sa/4.0/)  
-2024 Víctor A. Bettachini
-
-<!-- Remove for github pages -->
-<!-- Access to the material: [https://unlam.github.io/ComputationalAnalyticalMechanics/](https://unlam.github.io/ComputationalAnalyticalMechanics/) -->
-
-
-This is the repository for the course [Computational Analytical Mechanics](https://ingenieria.unlam.edu.ar/preview.php?seccion=3&idArticulo=510), for the [mechanical engineering degree](https://ingenieria.unlam.edu.ar/index.php?seccion=3&idArticulo=371) at the [Department of Engineering and Research in Technology](https://ingenieria.unlam.edu.ar/), Universidad Nacional de La Matanza.
-
-It's a code-based undergraduate course on analytical mechanics for engineering students with little to no prior programming knowledge. This 16-week flipped classroom course focuses on providing skills to calculate dynamics and strains of simple mechanical devices modelled as rigid bodies by solving Euler-Lagrange equations. Problems sets with increasingly complex exercises are introduced each week. Python-based solutions to previous ones are reused by students, who through small incremental modifications over the code presented by the teaching staff, build their own library of solutions to address mechanical modelling challenges.  
-
-The material for this course is available as Jupyter notebooks, containing theory and example exercises alongside the code to solve them.
-Students are required to rely on computer-based calculations, both analytical and numerical, so to focus their efforts on the physics modelling and interpretation of results. The online hosting and execution of these notebooks avoids any installation or hardware requirement issues.
-
-
-## Contents
-In this proposed schedule for a course of 16 weeks, the topics are divided by areas.
-Some of them are covered through more than one week, so a further descriptor, _Folder_, indicates where to find each week's material.
-
-| Week | Area                  | Folder            | Topics                   |
-|:----:|:---------------------:|:----------------:|:-------                |
-| 01   | Newtonian mechanics   | 01Vector           | Course presentation. Vector calculus using Sympy. |  
-| 02   | Analytical mechanics  | 02Energy           | Degrees of freedom. Generalized coordinates. Kinetic and potential energies. |
-| 03   | "                     | 03EulerLagrange    | Euler-Lagrange equations. |
[…]
