+++ docs/_config.yml
@@ -1,27 +1,0 @@
-title: Computational Analytical Mechanics
-author:
-  name: Victor Bettachini
-  email: "ask@for.it"
-logopath: assets/img/bothLogo.png
-logowidth: 100
-logoalt: Logo UNLAM
-logolink: https://ingenieria.unlam.edu.ar/
-description: >-
-  DIIT UNLAM
-# baseurl: /MecanicaAnaliticaComputacional
-baseurl: /ComputationalAnalyticalMechanics
-# url: https://bettachini.github.io/
-url: https://unlam.github.io/
-github_username: unlam
-# github_username: frautn
-
-# Build settings
-theme: minima
-
-
-# plugins:
-  # - jekyll-feed
-
-# header_pages:
-  # - about.md
-  # - portfolio.md
