+++ paper/back_paper.md
@@ -1,250 +1,0 @@
----
-title: 'Python-based Lagrange analytical mechanics course'
-tags:
-  - Python
-  - physics
-  - analitical mechanics
-  - mechanical engineering
-  - Lagrange
-  - flipped classroom
-authors:
-  - name: Víctor A. Bettachini
-    orcid: 0000-0001-7485-8884
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)
-  - name: Mariano A. Real
-    orcid: 0000-0003-3022-7516
-    affiliation: "3, 4"
-  - name: Edgardo Palazzo
-    orcid: 0009-0006-8783-8261
-    affiliation: "5"
-affiliations:
- - name: Universidad Nacional de La Matanza - UNLaM, Buenos Aires, Argentina.
-   index: 1
- - name: Instituto Geográfico Nacional - IGN, Buenos Aires, Argentina.
-   index: 2
- - name: Instituto Nacional de Tecnología Industrial - INTI, Buenos Aires, Argentina.
-   index: 3
- - name: INCALIN, Universidad de San Martín - UNSAM, Buenos Aires, Argentina.
-   index: 4
- - name: Universidad Tecnológica Nacional - UTN, Buenos Aires, Argentina.
-   index: 5
- 
-date: 02 September 2024
-bibliography: paper.bib
-lang: "en-GB"
----
-
-<!--  Notas sobre modificaciones y cambios realizados
-1. single-degree, single degree: ver nota de estilo de PR. También libro Fasano. Lo usaría como "single degree".
-2. "constraint forces", en Fasano, et al aparece como "constraint reaction" y luego lo llaman fuerza (p. 77)
--->
-
-
-
-# Summary
-We present a code-based undergraduate course on analytical mechanics for engineering students with little to no prior programming knowledge.
-This 16-week flipped classroom [@bishop_flipped_2013:2013] course provides skills to calculate dynamics and strains of simple mechanical devices, modelled as rigid bodies by solving Euler-Lagrange equations.
-<!--  Modificado MAR_20240830
-Each example and practice exercise are solved using computer-based analytical and numerical calculations with the aim to deflect students' focus of attention away from repetitive mathematical tasks towards the physics modelling.
--->
-Each example and practice exercise is solved using computer-based analytical and numerical calculations focusing students' attention on physics modelling and not on repetitive mathematical tasks.
-This approach also aims to improve creativity, the students have to solve problems by trial and error [@hoffmann_creatividad_2021:2021].
-
-<!--  Agregé la frase anterior en función del siguiente trabajo:
-Cita de freiberg2021creatividad:2021
-
-Para mejorar el rendimiento académico se necesita aplicar estrategias de enseñanza y evaluación que desestimulen el aprendizaje Superficial y promuevan el proceso creativo en los estudiantes. Los cambios en las prácticas educativas deberán, además, enfocarse en mejorar el aprendizaje Profundo y la creatividad para ajustar el perfil profesional de habilidades de los egresados al requerido actualmente por el ámbito laboral.
-
-De esta manera, el aprendizaje Profundo se encuentra estrechamente relacionado con un tipo de enseñanza basada en su conocimiento previo, la puesta en práctica a través de pruebas de ensayo y error, y la evaluación holística de los contenidos, entre otras cosas (Biggs & Tang, 2011).
--->
-
-The course addresses specific regional issues faced by third-year Latin American students (mid-career), who by then have learned how to solve ordinary differential equations.
-Theory and examples exercises, along with the _Python_ code that solves them are presented in _Jupyter notebooks_ run online to avoid installation and hardware requirement issues.
-Currently, the material is available in a GitHub repository in Spanish and has only been partially translated into English.
-<!--
[…]
