+++ paper/back_paper.bib
@@ -1,280 +1,0 @@
-@article{masonComparing:2013,
-	title = {Comparing the Effectiveness of an Inverted Classroom to a Traditional Classroom in an Upper-Division Engineering Course},
-	volume = {56},
-	issn = {0018-9359, 1557-9638},
-	url = {http://ieeexplore.ieee.org/document/6481483/},
-	doi = {10.1109/TE.2013.2249066},
-	abstract = {An inverted, or ﬂipped, classroom, where content delivery includes video lectures watched outside of the classroom, is a method that can free classroom time for learner-centered activities such as active and problem-based learning. This study compared the effectiveness of an inverted classroom to a traditional classroom in three areas: 1) content coverage; 2) student performance on traditional quizzes and exam problems; and 3) student observations and perception of the inverted classroom format. A controltreatment experiment comparing an inverted classroom to a traditional lecture-style format was used. The results show that: 1) the inverted classroom allowed the instructor to cover more material; 2) students participating in the inverted classroom performed as well or better on comparable quiz and exam questions and on openended design problems; and 3) while students initially struggled with the new format, they adapted quickly and found the inverted classroom format to be satisfactory and effective.},
-	pages = {430--435},
-	number = {4},
-	journaltitle = {{IEEE} Transactions on Education},
-	shortjournal = {{IEEE} Trans. Educ.},
-	author = {Mason, Gregory S. and Shuman, Teodora Rutar and Cook, Kathleen E.},
-	urldate = {2023-03-13},
-	date = {2013-11},
-	langid = {english},
-	file = {Mason et al. - 2013 - Comparing the Effectiveness of an Inverted Classro.pdf:/home/vbettachini/storage/Zotero/storage/JD5VKTX8/Mason et al. - 2013 - Comparing the Effectiveness of an Inverted Classro.pdf:application/pdf},
-}
-
-
-
-@article{cumbyCourse:2023,
-	title = {Course Materials for an Introduction to Data-Driven Chemistry},
-	volume = {6},
-	issn = {2577-3569},
-	url = {https://jose.theoj.org/papers/10.21105/jose.00192},
-	doi = {10.21105/jose.00192},
-	abstract = {Cumby et al., (2023). Course Materials for an Introduction to Data-Driven Chemistry. Journal of Open Source Education, 6(63), 192, https://doi.org/10.21105/jose.00192},
-	pages = {192},
-	number = {63},
-	journaltitle = {Journal of Open Source Education},
-	author = {Cumby, James and Degiacomi, Matteo T. and Erastova, Valentina and Güven, J. Jasmin and Hobday, Claire L. and Mey, Antonia S. j s and Pollak, Hannah and Szabla, Rafal},
-	urldate = {2024-07-30},
-	date = {2023-05-20},
-	langid = {english},
-	file = {Full Text PDF:/home/vbettachini/storage/Zotero/storage/HQZH5ADM/Cumby et al. - 2023 - Course Materials for an Introduction to Data-Drive.pdf:application/pdf},
-}
-
-@online{noauthorLagranges,
-	title = {Lagrange’s Method in Physics/Mechanics - {SymPy} 1.13.2 documentation},
-	url = {https://docs.sympy.org/latest/modules/physics/mechanics/lagrange.html},
-	urldate = {2024-08-19},
-	file = {Lagrange’s Method in Physics/Mechanics - SymPy 1.13.2 documentation:/home/vbettachini/storage/Zotero/storage/TCFNB4YU/lagrange.html:text/html},
-}
-
-@article{         harrisArrayNumpy:2020,
- title         = {Array programming with {NumPy}},
- author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
-                 van der Walt and Ralf Gommers and Pauli Virtanen and David
[…]
