### 2025-10-28 bcfbf0e
README.md contact and reference update
