### 2025-10-29 f9ec4e7
README w/ colab badges
