+++ old_README.md
@@ -1,0 +1,261 @@
+<!-- Remove for github pages -->
+[<image src="docs/assets/img/bothLogo.svg" alt="Logo UNLAM"  width="200"/>](https://ingenieria.unlam.edu.ar/)
+
+<!-- Uncomment for github pages -->
+<!-- [<image src="assets/img/bothLogo.svg" alt="Logo UNLAM"  width="200"/>](https://ingenieria.unlam.edu.ar/) -->
+
+# Computational Analytical Mechanics
+
+<!-- Remove for github pages -->
+[<image src="https://upload.wikimedia.org/wikipedia/commons/1/12/Cc-by-nc-sa_icon.svg"  width="80"/>](https://creativecommons.org/licenses/by-nc-sa/4.0/) 2025 Víctor A. Bettachini
+
+<!-- Uncomment for github pages -->
+<!-- [<image src="https://upload.wikimedia.org/wikipedia/commons/1/12/Cc-by-nc-sa_icon.svg"  width="80"/>](https://creativecommons.org/licenses/by-nc-sa/4.0/)   -->
+<!-- 2024 Víctor A. Bettachini -->
+
+<!-- Remove for github pages -->
+<!--
+Access to the material: [https://unlam.github.io/ComputationalAnalyticalMechanics/](https://unlam.github.io/ComputationalAnalyticalMechanics/)
+--->
+
+This is the repository for the course [Computational Analytical Mechanics](https://ingenieria.unlam.edu.ar/preview.php?seccion=3&idArticulo=510), for the [mechanical engineering degree](https://ingenieria.unlam.edu.ar/index.php?seccion=3&idArticulo=371) at the [Department of Engineering and Research in Technology](https://ingenieria.unlam.edu.ar/), Universidad Nacional de La Matanza.
+A [Spanish version](https://github.com/unlam/MecanicaAnaliticaComputacional) is available.
+
+It's a code-based undergraduate course on analytical mechanics for engineering students with little to no prior programming knowledge.
+This 16-week flipped classroom course focuses on providing skills to calculate dynamics and strains of simple mechanical devices modelled as rigid bodies by solving Euler-Lagrange equations.
+Problems sets with increasingly complex exercises are introduced each week.
+Python-based solutions to previous ones are reused by students, who through small incremental modifications over the code presented by the teaching staff, build their own library of solutions to address mechanical modelling challenges.  
+
+The material for this course is available as [Jupyter notebooks](https://jupyter.org/), containing theory and example exercises alongside the code to solve them.
+Students are required to rely on computer-based calculations, both analytical and numerical, so to focus their efforts on the physics modelling and interpretation of results.
+The online hosting and execution of these notebooks avoids any installation or hardware requirement issues.
+
+
+## Requirements
+Notebooks in this repository require to explicitly import the following Python libraries:
+- [SymPy](https://www.sympy.org/en/index.html) 1.14.0
+- [NumPy](https://numpy.org/) 2.3.3
+- [SciPy](https://scipy.org/) 1.16.2
+- [Matplotlib](https://matplotlib.org/) 3.10.6
+
+
+## Get in touch
+If you have a question regarding this course, feel free to [open an issue](https://github.com/unlam/ComputationalAnalyticalMechanics/issues) on its GitHub repository.
+We welcome community feedback and suggestions!  
+
+
+## Schedule
+In this flipped classroom course, a problem set has to be turned in by the students at the end of synchronous weekly meeting.
[…]
