### 2025-10-29 7ec0669
simplified README
