+++ README.md
@@ -1,23 +1,16 @@
+---
+lang: en-GB
+---
+
 # Computational Analytical Mechanics
 
-This is the repository for the [Computational Analytical Mechanics](https://ingenieria.unlam.edu.ar/preview.php?seccion=3&idArticulo=510) course for the [mechanical engineering degree](https://ingenieria.unlam.edu.ar/index.php?seccion=3&idArticulo=371) at the [Department of Engineering and Research in Technology](https://ingenieria.unlam.edu.ar/), Universidad Nacional de La Matanza.
+All teaching material used in the [Computational Analytical Mechanics](https://ingenieria.unlam.edu.ar/preview.php?seccion=3&idArticulo=510) course for the [mechanical engineering degree](https://ingenieria.unlam.edu.ar/index.php?seccion=3&idArticulo=371) at the Universidad Nacional de La Matanza is published here.
 A [Spanish version](https://github.com/unlam/MecanicaAnaliticaComputacional) is also available.
 
-It's a code-based undergraduate course on analytical mechanics for engineering students with little to no prior programming knowledge.
-This 16-week flipped classroom course focuses on providing skills to calculate dynamics and strains of simple mechanical devices modelled as rigid bodies by solving Euler-Lagrange equations.
-Problems sets with increasingly complex exercises are introduced each week.
-[Python](https://www.python.org/)-based solutions to previous ones are reused by students, who through small incremental modifications over the code presented by the teaching staff, build their own library of solutions to address mechanical modelling challenges.  
-
-The material for this course is available as [Jupyter notebooks](https://jupyter.org/), containing theory and example exercises alongside the code to solve them.
-Students are required to rely on computer-based calculations, both analytical and numerical, so to focus their efforts on the physics modelling and interpretation of results.
-The online hosting and execution of these notebooks avoids any installation or hardware requirement issues.
-
-Problem sets are provided as PDF files.
-Their [$\LaTeX$](https://www.latex-project.org/) source code is provided in order for instructors to modify them at please.
-
 <!-- Remove for github pages -->
-[<image src="figuresLaTeX/ambos.svg" alt="Logo UNLAM"  width="150"/>](https://ingenieria.unlam.edu.ar/)
-[<image src="https://upload.wikimedia.org/wikipedia/commons/1/12/Cc-by-nc-sa_icon.svg"  width="80"/>](https://creativecommons.org/licenses/by-nc-sa/4.0/) 2025 Víctor A. Bettachini
+[![DIIT-UNLaM](figuresLaTeX/ingenieria_logo_schwarz.png)](https://ingenieria.unlam.edu.ar/)
+[![CC BY-NC-SA](https://upload.wikimedia.org/wikipedia/commons/1/12/Cc-by-nc-sa_icon.svg)](https://creativecommons.org/licenses/by-nc-sa/4.0/)
+ 2025 Víctor A. Bettachini
 <!-- Uncomment for github pages -->
 <!-- [<image src="assets/img/bothLogo.svg" alt="Logo UNLAM"  width="200"/>](https://ingenieria.unlam.edu.ar/) -->
 <!-- Remove for github pages -->
@@ -31,14 +24,58 @@ Their [$\LaTeX$](https://www.latex-project.org/) source code is provided in orde
 Access to the material: [https://unlam.github.io/ComputationalAnalyticalMechanics/](https://unlam.github.io/ComputationalAnalyticalMechanics/)
 --->
 
+## Course Overview
+
+This undergraduate course introduces analytical mechanics through computational methods, designed for engineering students with little to no programming background. Students learn to model mechanical devices as rigid bodies and solve Euler-Lagrange equations to analyse dynamics and strains.
+
+### Learning Materials
+
[…]
