+++ README.md
@@ -1,6 +1,4 @@
----
-lang: en-GB
----
+<!-- LTeX: language=en-GB -->
 
 # Computational Analytical Mechanics
 
@@ -27,6 +25,7 @@ Access to the material: [https://unlam.github.io/ComputationalAnalyticalMechanic
 ## Course Overview
 
 This undergraduate course introduces analytical mechanics through computational methods, designed for engineering students with little to no programming background. Students learn to model mechanical devices as rigid bodies and solve Euler-Lagrange equations to analyse dynamics and strains.
+
 
 ### Learning Materials
 
