### 2025-10-29 f491aba
LaTeX README.md
