### 2025-10-30 cf13b6b
README.md methodology
