+++ README.md
@@ -95,7 +95,7 @@ Some of them are covered through more than one week, so a further descriptor, _F
 
 | Week | Area                 | Folder             | Topics |
 |:-----|:--------------------:|:------------------:|:-------|
-| 01   | Newtonian mechanics  | 01Vector           | Course presentation. Vector calculus using SymPy. |  
+| 01   | Newtonian mechanics  | 01Vector           | Course methodology. Vector calculus using SymPy. |  
 | 02   | Analytical mechanics | 02Energy           | Degrees of freedom. Generalized coordinates. Kinetic and potential energies. |
 | 03   | "                    | 03EulerLagrange    | Euler-Lagrange equations. |
 | 04   | "                    | 04Constraints      | Constraints as functions of coordinates. |
