+++ README.md
@@ -1,8 +1,7 @@
 <!-- LTeX: language=en-US -->
 
 # Computational Analytical Mechanics
-
-All teaching material used in the [Computational Analytical Mechanics](https://ingenieria.unlam.edu.ar/preview.php?seccion=3&idArticulo=510) course for the [mechanical engineering degree](https://ingenieria.unlam.edu.ar/index.php?seccion=3&idArticulo=371) at the Universidad Nacional de La Matanza is published here.
+Here it is published all teaching material used in the [course on this subject](https://ingenieria.unlam.edu.ar/preview.php?seccion=3&idArticulo=510) for the [mechanical engineering degree](https://ingenieria.unlam.edu.ar/index.php?seccion=3&idArticulo=371) at the Universidad Nacional de La Matanza.
 A [Spanish version](https://github.com/unlam/MecanicaAnaliticaComputacional) is also available.
 
 <!-- Remove for github pages -->
@@ -105,7 +104,7 @@ Some of them are covered through more than one week, so a further descriptor, _F
 | 08InertiaTensor    | Rigid body           | Inertia tensor of point masses systems. Steiner theorem. |
 | 09DistributedMass  | "                    | Inertia tensor of masses distributions. |
 | 10EulerRotation    | "                    | Euler equations for the rigid body. |
-| 11 (continues 10)  | "                    | **Final project: statement of the problem** |
+| 11 (continues 10)  | "                    | **Final project: discussion on the statement of the problem** |
 | 12OscillationsSDOF | Oscillations         | Forced oscillations in single degree of freedom systems. |
 | 13OscillationsMDOF | "                    | Forced oscillations in multiple degrees of freedom systems. |
 | 14 (continues 13)  | "                    | " |
@@ -139,7 +138,7 @@ https://img.shields.io/badge/Google%20Colab-000000?logo=googlecolab&logoColor=ff
 ### 02 Kinetic energy and gravitational potential energy
 - [Kinetic energy](https://www.github.com/unlam/ComputationalAnalyticalMechanics/blob/main/02Energy/kineticEnergy.ipynb) [![Kinetic energy](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/unlam/ComputationalAnalyticalMechanics/blob/main/02Energy/kineticEnergy.ipynb)
 - [Gravitational potential energy: pendulum with pivot free to slide horizontally](https://www.github.com/unlam/ComputationalAnalyticalMechanics/blob/main/02Energy/gravitationalPotentialEnergy.ipynb) [![Gravitational potential energy: pendulum with pivot free to slide horizontally](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/unlam/ComputationalAnalyticalMechanics/blob/main/02Energy/gravitationalPotentialEnergy.ipynb)
-- [Energy problem set](https://github.com/unlam/ComputationalAnalyticalMechanics/blob/main/02Energy/exercisesEnergy.pdf)
+- [Problem set - Energy](https://github.com/unlam/ComputationalAnalyticalMechanics/blob/main/02Energy/exercisesEnergy.pdf)
   - pset02e02
   - pset02e03
   - pset02e04
@@ -149,21 +148,19 @@ https://img.shields.io/badge/Google%20Colab-000000?logo=googlecolab&logoColor=ff
[…]
