+++ README.md
@@ -122,23 +122,22 @@ https://img.shields.io/badge/Google%20Colab-000000?logo=googlecolab&logoColor=ff
 
 ### 01 Vector kinematics
 
-- [Introduction to this course](https://www.github.com/unlam/ComputationalAnalyticalMechanics/blob/main/01Vector/thisCourse.ipynb)
-[![Introduction to this course](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/unlam/ComputationalAnalyticalMechanics/blob/main/01Vector/thisCourse.ipynb)
+- [This course subject matter](https://www.github.com/unlam/ComputationalAnalyticalMechanics/blob/main/reference/academical/subjectMatter.ipynb)
+[![Introduction to this course](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/unlam/ComputationalAnalyticalMechanics/blob/main/reference/academical/subjectMatter.ipynb)
 - [Vector kinematics](https://www.github.com/unlam/ComputationalAnalyticalMechanics/blob/main/01Vector/vectorKinematics.ipynb)
 [![Vector kinematis](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/unlam/ComputationalAnalyticalMechanics/blob/main/01Vector/vectorKinematics.ipynb)
-- [First problem set (pset01) - Vector kinematics](https://github.com/unlam/ComputationalAnalyticalMechanics/blob/main/01Vector/exercisesVector.pdf) | Exercises to turn-in:
-  - pset01e01 one hour after being presented. 
+- [First problem set (pset01) - Vector kinematics](https://github.com/unlam/ComputationalAnalyticalMechanics/blob/main/01Vector/exercisesVector.pdf)
+[![First problem set (pset01) - Vector kinematics](https://img.shields.io/badge/-PDF-red?logo=adobeacrobatreader&logoColor=white)](https://github.com/unlam/ComputationalAnalyticalMechanics/raw/refs/heads/main/01Vector/exercisesVector.pdf)
+  - Exercise pset01e01 should be turned-in in one hour.
   - pset01e02 at next week meeting beginning.
-  - Those of pset02 at end of next meeting (should be started during the week). 
-- References
-  - [Introduction to Python](https://www.github.com/unlam/ComputationalAnalyticalMechanics/blob/main/01Vector/IntroPython.ipynb) [![Introduction to Python](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/unlam/ComputationalAnalyticalMechanics/blob/main/01Vector/IntroPython.ipynb)
-  - [Markdown and LaTeX](https://www.github.com/unlam/ComputationalAnalyticalMechanics/blob/main/01Vector/MarkdownLaTeX.ipynb) [![Markdown and LaTeX](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/unlam/ComputationalAnalyticalMechanics/blob/main/01Vector/MarkdownLaTeX.ipynb)
+  - Those of pset02 at end of next meeting (begin during the week). 
 
 
 ### 02 Kinetic energy and gravitational potential energy
 - [Kinetic energy](https://www.github.com/unlam/ComputationalAnalyticalMechanics/blob/main/02Energy/kineticEnergy.ipynb) [![Kinetic energy](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/unlam/ComputationalAnalyticalMechanics/blob/main/02Energy/kineticEnergy.ipynb)
[…]
