+++ 01Vector/exercisesVector.tex
@@ -4,7 +4,7 @@
 
 \begin{document}
 \begin{center}
-	\textsc{\large Position vector}
+	\textsc{\large Vectorial kinematics}
 \end{center}
 
 If you are able to solve these problems on your own, then you can assume that you have the minimum knowledge about these topics. 
