+++ 01Vector/exercisesVector.tex
@@ -1,3 +1,4 @@
+% LTeX: language=en-US
 \documentclass[11pt, a4paper, twoside]{article}
 
 \input{../figuresLaTeX/introUNLaM}
@@ -13,7 +14,7 @@ If you are able to solve these problems on your own, then you can assume that yo
 The problems marked with (*) have additional difficulties. Don't hesitate about seeking help from teachers and your classmates if you are not able to complete them.
 		
 \noindent
-For each of the following excercises, create a Jupyter notebook with your name in the title, including one or several cells of code interwined with cells showing text indicating the excercise that is being solved.
+For each of the following exercises, create a Jupyter notebook with your name in the title, including one or several cells of code intertwined with cells showing text indicating the exercise that is being solved.
 
 \begin{enumerate}
 % 	\section*{Vector posición}
@@ -88,7 +89,7 @@ For each of the following excercises, create a Jupyter notebook with your name i
 			(*) \textbf{Position as a function of a variable}\\
 			A particle of mass \(m\) is attached to a ring of radius $R$, and therefore its radius measured from the center of the ring is constant. Then it's enough to know the angle $\varphi$ to describe its position.
 			\begin{enumerate}
-				\item Write it using cartesian coordinates in terms of $R$ and $\varphi$. Recall that $R$ is a constant, it's just a symbol for \verb'SymPy', and $\varphi$ is a variable that depends on time, or dynamic symbol, as it is called in this library. You will need to use trigonometric functions, investigate how these are implemented using this library.
+				\item Write it using Cartesian coordinates in terms of $R$ and $\varphi$. Recall that $R$ is a constant, it's just a symbol for \verb'SymPy', and $\varphi$ is a variable that depends on time, or dynamic symbol, as it is called in this library. You will need to use trigonometric functions, investigate how these are implemented using this library.
 				\item Calculate the velocity of this particle using \verb'SymPy'.\\
 				Answer:\\ 
 				\(
