+++ 02Energy/exercisesEnergy.tex
@@ -1,3 +1,4 @@
+% LTeX: language=en-US
 \documentclass[11pt, a4paper, twoside]{article}
 
 \input{../figuresLaTeX/introUNLaM}
@@ -34,7 +35,7 @@ Exercises marked with (*) have extra difficulty, don't hesitate to ask for help.
 \item
 	\begin{minipage}[t][3.7cm]{0.7\textwidth}
 		\textbf{Double coplanar pendulum} [Landau \S5 ex. 1]\\
-		A ridig bar of lentgh $\ell_1$ and negligible mass has a particle of mass $m_1$ attached to one end. There is a second bar of negligible mass hanging from the first one, of length $\ell_2$, with a particle of mass $m_2$ attached to the other end too.		
+		A rigid bar of length $\ell_1$ and negligible mass has a particle of mass $m_1$ attached to one end. There is a second bar of negligible mass hanging from the first one, of length $\ell_2$, with a particle of mass $m_2$ attached to the other end too.		
 	\end{minipage}
 	\begin{minipage}[c][2cm][t]{0.3\textwidth}
 		\includegraphics[width=0.75\textwidth]{figures/landauS52_fig1.png}
@@ -56,7 +57,7 @@ Exercises marked with (*) have extra difficulty, don't hesitate to ask for help.
 \item
 	\begin{minipage}[t][7.1cm]{0.5\textwidth}
 		(*) \textbf{Pendulum with rotating point of support} [Marion (e) ex. 7.5] [Landau \S5 ex. 3]\\
-		A particle of mass $m$ is attached to the end of a rigid bar of length $b$. The point of support is linked to a vertical circle of radius $a$ and it rotates with constant frequency $\omega$. It is assumed that all positions lie in the same plane and the mass of the bar is negligible. Calculate the kinetic energy, $T$, and potential $V$, of the particle of mass $m$.
+		A particle of mass $m$ is attached to the end of a rigid bar of length $b$. The point of support is linked to a vertical circle of radius $a$, and it rotates with constant frequency $\omega$. It is assumed that all positions lie in the same plane and the mass of the bar is negligible. Calculate the kinetic energy, $T$, and potential $V$, of the particle of mass $m$.
 
 		Result:\\
 		\(
@@ -91,7 +92,7 @@ Exercises marked with (*) have extra difficulty, don't hesitate to ask for help.
 		\includegraphics[width=0.35\textwidth]{figures/rotantesAcopladas2.png}
 	\end{center}
 
-	In these figures, the top sphere is the fixed point \(A\). Everything revolves around the pink axis with constant angular speed $\Omega$. Therefore, the particles on each side, of mass $m_1$, rotate entering and leaving the plane shown in the first image. This is equivalent to thinking that the lightblue plane rotates around the pink axis.
+	In these figures, the top sphere is the fixed point \(A\). Everything revolves around the pink axis with constant angular speed $\Omega$. Therefore, the particles on each side, of mass $m_1$, rotate entering and leaving the plane shown in the first image. This is equivalent to thinking that the light-blue plane rotates around the pink axis.
 
 	The piece of mass $m_2$ is a sliding hub that moves vertically without friction. 
 	
