### 2025-11-02 19b0a86
10EulerRotation/e05 error
