+++ 10EulerRotation/exercisesEulerRotation.tex
@@ -78,7 +78,7 @@
 		\begin{tasks} 
 			\task Calculate the force applied by the bearings.\\
 			Result: \(
-				F = \frac{m a^2 \omega^2}{D} \sin(\varphi) \cos(\varphi)
+				F = \frac{2 m a^2 \omega^2}{D} \sin(\varphi) \cos(\varphi)
 			\).
 		\end{tasks}
 	\end{minipage}
