### 2025-11-02 49e90ca
10EulerRotation/exercisesEulerRotation e
