+++ 10EulerRotation/exercisesEulerRotation.tex
@@ -110,9 +110,9 @@
 		\begin{tasks}
 			\task Convert the data in imperial units (feet, pounds) to International System units.
 			\task Determine the reactions that the bearings must provide.\\
-			Result: \(
-			\left[\begin{matrix}A_{y}\\A_{z}\\B_{x}\\B_{y}\\B_{z}\end{matrix}\right] = \left[\begin{matrix}-5.79\\21.1 t + 33.4\\0\\5.79\\33.4 - 21.1 t\end{matrix}\right]
-			\)
+%			Result: \(
+%			\left[\begin{matrix}A_{y}\\A_{z}\\B_{x}\\B_{y}\\B_{z}\end{matrix}\right] = \left[\begin{matrix}-5.79\\21.1 t + 33.4\\0\\5.79\\33.4 - 21.1 t\end{matrix}\right]
+%			\)
 		\end{tasks}
 	\end{minipage}
 	\begin{minipage}[c][2cm][t]{0.3\textwidth}
