### 2025-11-03 766d19b
10EulerRotation/exercisesEulerRotation
