### 2025-11-05 bc9f9f8
10EulerRotation/exercisesEulerRotation
