+++ 10EulerRotation/exercisesEulerRotation.tex
@@ -117,7 +117,7 @@
 		\task Convert the data in imperial units (feet, pounds) to International System units.
 		\task Determine the reactions that the bearings must provide.\\
 		Result: \(
-		\left[\begin{matrix}A_{y}\\A_{z}\\B_{x}\\B_{y}\\B_{z}\end{matrix}\right] = \left[\begin{matrix}-17.4\\33.4\\134.0\\17.4\\33.4\end{matrix}\right]
+			\left[\begin{matrix}A_{y}\\A_{z}\\B_{x}\\B_{y}\\B_{z}\end{matrix}\right] = \left[\begin{matrix}-7.52\\33.4\\268.0\\-17.4\\33.4\end{matrix}\right]
 		\)
 	\end{tasks}
 
