### 2025-11-24 #293 @emckiernan
Hi, @ata0k, thank you for accepting my invitation to review! I'll tag you below to formally assign you. Here are the reviewer guidelines: https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html. If you have any questions, please let me know. Thanks again!
