### 2025-11-24 #293 @Carreau
> Hi, [@Carreau](https://github.com/Carreau). I'm reaching out because you previously signed up to review for JOSE. We're looking for another reviewer for this submission. Would you be willing to contribute a review? Please let me know if you're willing and available. Thanks in advance for considering!

Sorry it took me a while to answer; at the moment I'm a bit oversubscribed, so I'll decline for the time being.
