### 2025-12-02 #293 @labarba
@ata0k – do run the command `@editorialbot generate my checklist` and it should generate your reviewer checklist (if all goes well)
