+++ README.md
@@ -1,9 +1,9 @@
 <!-- LTeX: language=en-US -->
 
 # Computational Analytical Mechanics
-Here it is published all teaching material used in the [course on this subject](https://ingenieria.unlam.edu.ar/preview.php?seccion=3&idArticulo=510) for the [mechanical engineering degree](https://ingenieria.unlam.edu.ar/index.php?seccion=3&idArticulo=371) at the Universidad Nacional de La Matanza.
+This repository hosts all the teaching materials for the [introductory course to Analytical Mechanics](https://ingenieria.unlam.edu.ar/preview.php?seccion=3&idArticulo=510) of the [Mechanical Engineering degree](https://ingenieria.unlam.edu.ar/index.php?seccion=3&idArticulo=371) at the Universidad Nacional de La Matanza.
 
-Una versión en castellano de este repositorio, [Mecánica Analítica Computacional](https://github.com/unlam/MecanicaAnaliticaComputacional), se encuentra disponible.
+Una versión en castellano de este repositorio, [Mecánica Analítica Computacional](https://github.com/unlam/MecanicaAnaliticaComputacional), también está disponible.
 
 
 <!-- Remove for github pages -->
@@ -23,22 +23,47 @@ Una versión en castellano de este repositorio, [Mecánica Analítica Computacio
 Access to the material: [https://unlam.github.io/ComputationalAnalyticalMechanics/](https://unlam.github.io/ComputationalAnalyticalMechanics/)
 --->
 
-## Course Overview
+## Table of Contents
+- [Abount the Course](#about-the-course)
+  - [Learning Materials](#learning-materials)
+  - [Teaching Approach](#teaching-approach)
+  - [Technical Features](#technical-features)
+  - [Get in touch](#get-in-touch)
+- [Weekly Schedule](#weekly-schedule)
+  - [01 Vector kinematics](#01-vector-kinematics)
+  - [02 Kinetic energy and gravitational potential energy](#02-kinetic-energy-and-gravitational-potential-energy)
+  - [03 Euler-Lagrange equation](#03-euler-lagrange-equation)
+  - [04 Constraints](#04-constraints)
+  - [05 Numerical simulations](#05-numerical-simulations)
+  - [06 Constraint reactions](#06-constraint-reactions)
+  - [07 Nonconservative forces](#07-nonconservative-forces)
+  - [08 The inertia tensor](#08-the-inertia-tensor)
+  - [09 Distributed mass](#09-distributed-mass)
+  - [10 Euler equations for the rotation of rigid bodies](#10-euler-equations-for-the-rotation-of-rigid-bodies)
+  - [12 Oscillations of single degree of freedom systems](#12-oscillations-of-single-degree-of-freedom-systems)
+  - [13 Oscillations in multiple degrees of freedom systems](#13-oscillations-in-multiple-degrees-of-freedom-systems)
+  - [15 Final project](#15-final-project)
+- [Bibliography](#bibliography)
 
-This undergraduate course introduces analytical mechanics through computational methods.
-It has been designed for engineering students with little to no programming background.
-Students learn to model mechanical devices as rigid bodies, and through the use of the Euler-Lagrange equations, analyze their dynamics and strains.
+
+## About the Course
+
+- It introduces analytical mechanics through computational methods in the time frame of a four-month term.
+- Designed for undergrad engineering students with little to no programming background.
+- Mechanical devices as modelled as rigid bodies.
+- Device's dynamics and strains differential equations are derived using the Euler-Lagrange formalism.
+- All analytical and numerical calculations are computationally performed.
 
 
 ### Learning Materials
 
-New topics are presented at one or more **[Jupyter notebooks](https://jupyter.org/)** that combine:
[…]
