### 2026-04-08 #293 @ata0k
Here is my review report. I hope it will help to improve the paper and repositories.

**Review Report**

The repository contains many interesting problems that can help students acquire the necessary knowledge for an Analytical Mechanics course. Implementing these problems in Python notebooks is useful not only for understanding the problems themselves but also for approaching solutions from a computational perspective—the most common tool for solving theoretical problems in modern Physics. Furthermore, these notebooks aid in visualization. While this work will certainly be appreciated by teachers and students, there are several major pedagogical issues and minor errors (mostly typos) that must be addressed before the paper can be accepted.

**Major Pedagogical and Structural Issues**

**Motivation and Context:** Since the repository is intended for a flipped classroom methodology, it should function as an interactive book. Each topic requires a proper introduction, motivation, and an explanation of the formulas. For instance, the first section, "Position in a cartesian frame of reference," starts abruptly with a figure and an equation without explaining the objective or the derivation. By the way, it is Cartesian, with a capital C.

**Pedagogical Flow:** Several sections seem to start with exercises rather than theory. In the "Energy" topic, the first notebook mentions a "Pendulum with free support, again," despite it being the first time a pendulum is mentioned in a notebook of the repository. This lack of sequencing contradicts the proposed teaching methodology.

**Interactivity and Visualization:** One of the main advantages of Python notebooks is the possibility to create interactive visualizations (e.g., using sliders to change parameters). There is a noticeable lack of this; the first visualization does not appear until Topic 5, and the second one only in Topic 12. Early topics like Linear Algebra and Vectors would benefit significantly from interactive plots.

**Consistency and Organization:**

- In the English version, the "flywheel" notebook (Topic 10) is empty and is also missing from the Spanish version. Is it important?

-There are discrepancies in the numbering: the paper mentions 12 folders and 11 topics, while the repository is enumerated up to 15 (with only 13 actually present).
The English and Spanish versions are inconsistent in both numbering and content. I suggest that for notebooks not yet translated, the Spanish version should be included as a placeholder to maintain a consistent structure across both repositories.

**Minor Corrections and Technical Details**

**Language and Style:** There are several typos and grammatical issues throughout the repository; the language is often too informal for an academic resource in both repositories, but clearly more often in the English repository. I give some examples of this problem below, but there are many others. Please check the whole repository in each version.

**Figure Quality:** The quality and style of the figures are inconsistent. Many appear to be imported from external websites without proper citation. Ideally, figures should be generated directly in Python to ensure uniform quality and style.

**UI/UX:** Many cells have side-scrollbars that make navigation difficult, especially on tablets. These should be removed if possible.

**Specific Typos and Improvements**

**01 Vectors:**

- The first equation has a typo. $\vec{r} = r\hat{r}$, instead of $\vec{r} = r\vec{r}$.
It is confusing the notation. Sometimes they use $x,y,z$, sometimes $r_1,r_2,r_3$, sometimes $r_i,r_j,r_k$, sometimes $r_x, r_y,r_z$. They should try to keep the same notations throughout the whole notebook.
- Use “$:=$” when is a definition, for example $\dot{r} := \frac{dr}{dt}$.
Change “What if” to “What happens if” and “not cartesian” to “non-Cartesian”.
There is a typo on cell 12. “and” inside an equation.

**02 Energy:**

-"Pendulum with free support, again…", it is the first time the Pendulum is mentioned in the course. It has been mentioned in the exercise, but usually we expect the course first, then the exercises. So, at least it should be referenced to the exercises.

- Change "Actually one only force" to "a single force".
- Change "Such is the repulsion... that we merge with it" to "This repulsion is what prevents the molecules in our hand from binding with the object and prevents us from merging with it."
- Change "considerable interaction" to "significant force".
- For some reason, “directed downward” is in a different font size.

**03 Euler Lagrange:**

- Change “A spring has $l_0$ natural length…” to  “A spring has natural length $l_0$…”. Also change “$l = l(t)$ the…” to “$l = l(t)$ is the…”.
