### 2026-04-08 #293 @bettachini
Thanks for your comments @ata0k . I'll start to address the points you make, but I'll admit that it would take me a while until having implemented a rewrite of the material. Once I had done than I'll send a notice to @emckiernan  here in a further comment.
